@@ -3,15 +3,14 @@ plugins {
 }
 
 android {
-    compileSdk 34
-    buildToolsVersion = "30.0.3"
+    compileSdk 35
 
     defaultConfig {
         applicationId "ttit.com.shuvo.ikglhrm"
         minSdkVersion 23
-        targetSdkVersion 34
-        versionCode 6
-        versionName "1.1.4"
+        targetSdkVersion 35
+        versionCode 7
+        versionName "1.2.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -23,21 +22,22 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
     }
     namespace 'ttit.com.shuvo.ikglhrm'
 }
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.3.0'
-    implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.appcompat:appcompat:1.5.0'
+    implementation 'com.google.android.material:material:1.12.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation files('libs/ojdbc14.jar')
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.google.android.gms:play-services-maps:17.0.1'
     implementation 'com.google.android.gms:play-services-location:21.0.1'
+    implementation 'androidx.activity:activity:1.10.1'
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
@@ -47,7 +47,7 @@ dependencies {
     implementation "android.arch.work:work-runtime:1.0.1"
     implementation 'com.github.Mamoon-Alhawamdeh:AmazingSpinner:1.1.0'
     implementation 'com.android.volley:volley:1.2.1'
-    implementation 'androidx.work:work-runtime:2.8.1'
+    implementation 'androidx.work:work-runtime:2.9.1'
 
     //
     implementation 'de.hdodenhof:circleimageview:3.1.0'
@@ -55,4 +55,17 @@ dependencies {
     annotationProcessor 'com.github.bumptech.glide:compiler:4.13.2'
     implementation 'com.intuit.sdp:sdp-android:1.1.0'
     implementation 'com.google.android.play:app-update:2.1.0'
+    implementation 'com.google.code.gson:gson:2.10.1'
+
+    // Camera Lib
+    implementation 'io.fotoapparat:fotoapparat:2.6.1'
+
+    // crop image
+    implementation 'com.github.yalantis:ucrop:2.2.10'
+
+    //restart
+    implementation 'com.jakewharton:process-phoenix:2.1.2'
+
+    //calendar view
+    implementation 'com.applandeo:material-calendar-view:1.9.2'
 }
