@@ -5,15 +5,12 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.os.Bundle;
 import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.TextView;
 
@@ -22,10 +19,14 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
 import ttit.com.shuvo.ikglhrm.R;
 import ttit.com.shuvo.ikglhrm.WaitProgress;
 
 import static ttit.com.shuvo.ikglhrm.Login.userInfoLists;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
 
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
@@ -37,63 +38,23 @@ public class JobDescription extends AppCompatActivity {
     TextView no_job;
     Button jobFinish;
     RecyclerView job_list;
-    public static JobAdapter jobAdapter;
+    JobAdapter jobAdapter;
     RecyclerView.LayoutManager layoutManager;
 
     public static ArrayList<JobDescDetails> jobDescDetails;
 
     WaitProgress waitProgress = new WaitProgress();
-//    private String message = null;
     private Boolean conn = false;
-//    private Boolean infoConnected = false;
     private Boolean connected = false;
 
-//    private Connection connection;
 
     String emp_id = "";
 
-//    @Override
-//    public void onWindowFocusChanged(boolean hasFocus) {
-//        super.onWindowFocusChanged(hasFocus);
-//        if (hasFocus) {
-//            hideSystemUI();
-//        }
-//    }
-//    private void hideSystemUI() {
-//        View decorView = getWindow().getDecorView();
-//        decorView.setSystemUiVisibility(
-//                View.SYSTEM_UI_FLAG_IMMERSIVE
-//                        // Set the content to appear under the system bars so that the
-//                        // content doesn't resize when the system bars hide and show.
-//                        | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-//                        | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-//                        | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-//                        // Hide the nav bar and status bar
-//                        | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-//                        | View.SYSTEM_UI_FLAG_FULLSCREEN);
-//    }
+    Logger logger = Logger.getLogger(JobDescription.class.getName());
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-//        getWindow().requestFeature(Window.FEATURE_NO_TITLE);
-//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-//            Window w = getWindow();
-//            //w.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
-//        }
-//        if (Build.VERSION.SDK_INT < 16) {
-//            getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-//                    WindowManager.LayoutParams.FLAG_FULLSCREEN);
-//        }
-//        View decorView = getWindow().getDecorView();
-//        // Hide the status bar.
-//        int uiOptions = View.SYSTEM_UI_FLAG_FULLSCREEN;
-//        decorView.setSystemUiVisibility(uiOptions);
-        Window window = getWindow();
-        window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
-        window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-        window.setStatusBarColor(ContextCompat.getColor(JobDescription.this,R.color.secondaryColor));
-
         setContentView(R.layout.activity_job_description);
 
         emp_id = userInfoLists.get(0).getEmp_id();
@@ -103,7 +64,6 @@ public class JobDescription extends AppCompatActivity {
 
         jobDescDetails = new ArrayList<>();
 
-//        new Check().execute();
         getJobDescription();
 
         job_list.setHasFixedSize(true);
@@ -114,161 +74,9 @@ public class JobDescription extends AppCompatActivity {
 
 
 
-        jobFinish.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
+        jobFinish.setOnClickListener(v -> finish());
 
     }
-
-//    public boolean isConnected() {
-//        boolean connected = false;
-//        boolean isMobile = false;
-//        try {
-//            ConnectivityManager cm = (ConnectivityManager) getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
-//            NetworkInfo nInfo = cm.getActiveNetworkInfo();
-//            connected = nInfo != null && nInfo.isAvailable() && nInfo.isConnected();
-//            return connected;
-//        } catch (Exception e) {
-//            Log.e("Connectivity Exception", e.getMessage());
-//        }
-//        return connected;
-//    }
-//
-//    public boolean isOnline() {
-//
-//        Runtime runtime = Runtime.getRuntime();
-//        try {
-//            Process ipProcess = runtime.exec("/system/bin/ping -c 1 8.8.8.8");
-//            int     exitValue = ipProcess.waitFor();
-//            return (exitValue == 0);
-//        }
-//        catch (IOException | InterruptedException e)          { e.printStackTrace(); }
-//
-//        return false;
-//    }
-
-//    public class Check extends AsyncTask<Void, Void, Void> {
-//
-//        @Override
-//        protected void onPreExecute() {
-//            super.onPreExecute();
-//
-//            waitProgress.show(getSupportFragmentManager(),"WaitBar");
-//            waitProgress.setCancelable(false);
-//        }
-//
-//        @Override
-//        protected Void doInBackground(Void... voids) {
-//            if (isConnected() && isOnline()) {
-//
-//                JobDetails();
-//                if (connected) {
-//                    conn = true;
-//                    message= "Internet Connected";
-//                }
-//
-//            } else {
-//                conn = false;
-//                message = "Not Connected";
-//            }
-//
-//            return null;
-//        }
-//
-//        @Override
-//        protected void onPostExecute(Void aVoid) {
-//            super.onPostExecute(aVoid);
-//
-//            waitProgress.dismiss();
-//            if (conn) {
-//
-//                jobAdapter = new JobAdapter(jobDescDetails, JobDescription.this);
-//                job_list.setAdapter(jobAdapter);
-//                if (jobDescDetails.size() == 0) {
-//                    no_job.setVisibility(View.VISIBLE);
-//                    job_list.setVisibility(View.GONE);
-//                } else {
-//                    no_job.setVisibility(View.GONE);
-//                    job_list.setVisibility(View.VISIBLE);
-//                }
-//
-//                conn = false;
-//                connected = false;
-//
-//
-//
-//            }
-//            else {
-//                Toast.makeText(getApplicationContext(), "No Internet Connection", Toast.LENGTH_SHORT).show();
-//                AlertDialog dialog = new AlertDialog.Builder(JobDescription.this)
-//                        .setMessage("Please Check Your Internet Connection")
-//                        .setPositiveButton("Retry", null)
-//                        .setNegativeButton("Cancel",null)
-//                        .show();
-//
-//                dialog.setCancelable(false);
-//                dialog.setCanceledOnTouchOutside(false);
-//                Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-//                positive.setOnClickListener(new View.OnClickListener() {
-//                    @Override
-//                    public void onClick(View v) {
-//
-//                        new Check().execute();
-//                        dialog.dismiss();
-//                    }
-//                });
-//                Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-//                negative.setOnClickListener(new View.OnClickListener() {
-//                    @Override
-//                    public void onClick(View v) {
-//
-//                        dialog.dismiss();
-//                        finish();
-//                    }
-//                });
-//            }
-//        }
-//    }
-
-//    public void JobDetails() {
-//        try {
-//            this.connection = createConnection();
-//            //    Toast.makeText(MainActivity.this, "Connected",Toast.LENGTH_SHORT).show();
-//
-//
-//            jobDescDetails = new ArrayList<>();
-//            Statement stmt = connection.createStatement();
-//
-//
-//
-//
-//            ResultSet rs=stmt.executeQuery("SELECT EJD_JSDD_NAME FROM EMP_JOB_DESCRIPTION WHERE EJD_EMP_ID ="+emp_id+"");
-//
-//
-//            int i = 0;
-//            while(rs.next()) {
-//
-//                i++;
-//                jobDescDetails.add(new JobDescDetails(String.valueOf(i),rs.getString(1)));
-//
-//            }
-//
-//
-//            connected = true;
-//
-//            connection.close();
-//
-//        }
-//        catch (Exception e) {
-//
-//            //   Toast.makeText(MainActivity.this, ""+e,Toast.LENGTH_LONG).show();
-//            Log.i("ERRRRR", e.getLocalizedMessage());
-//            e.printStackTrace();
-//        }
-//    }
 
     public void getJobDescription() {
         waitProgress.show(getSupportFragmentManager(),"WaitBar");
@@ -278,7 +86,7 @@ public class JobDescription extends AppCompatActivity {
 
         jobDescDetails = new ArrayList<>();
 
-        String jobDescUrl = "http://103.56.208.123:8001/apex/ttrams/emp_information/getJobDescription/"+emp_id+"";
+        String jobDescUrl = api_url_front + "emp_information/getJobDescription/"+emp_id;
 
         RequestQueue requestQueue = Volley.newRequestQueue(JobDescription.this);
 
@@ -304,13 +112,13 @@ public class JobDescription extends AppCompatActivity {
             }
             catch (JSONException e) {
                 connected = false;
-                e.printStackTrace();
+                logger.log(Level.WARNING, e.getMessage(), e);
                 updateLayout();
             }
         },error -> {
             conn = false;
             connected = false;
-            error.printStackTrace();
+            logger.log(Level.WARNING, error.getMessage(), error);
             updateLayout();
         });
 
@@ -323,7 +131,7 @@ public class JobDescription extends AppCompatActivity {
             if (connected) {
                 jobAdapter = new JobAdapter(jobDescDetails, JobDescription.this);
                 job_list.setAdapter(jobAdapter);
-                if (jobDescDetails.size() == 0) {
+                if (jobDescDetails.isEmpty()) {
                     no_job.setVisibility(View.VISIBLE);
                     job_list.setVisibility(View.GONE);
                 } else {
