@@ -1,5 +1,6 @@
 package ttit.com.shuvo.ikglhrm.MainPage;
 
+import androidx.activity.OnBackPressedCallback;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.cardview.widget.CardView;
@@ -7,11 +8,9 @@ import androidx.cardview.widget.CardView;
 import android.app.ActivityManager;
 import android.app.PendingIntent;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -170,7 +169,7 @@ public class MainMenu extends AppCompatActivity {
         welcomeText = findViewById(R.id.welcome_text_view_main_menu);
 
         int currentHour = Calendar.getInstance().get(Calendar.HOUR_OF_DAY);
-        String wt = "";
+        String wt;
         if (currentHour >= 4 && currentHour <= 11) {
             wt = "GOOD MORNING,";
         }
@@ -196,7 +195,7 @@ public class MainMenu extends AppCompatActivity {
             userImage.setImageResource(R.drawable.profile);
         }
 
-        if (userInfoLists.size() != 0) {
+        if (!userInfoLists.isEmpty()) {
             String firstname = userInfoLists.get(0).getUser_fname();
             String lastName = userInfoLists.get(0).getUser_lname();
             if (firstname == null) {
@@ -209,7 +208,7 @@ public class MainMenu extends AppCompatActivity {
             userName.setText(empFullName);
         }
 
-        if (userDesignations.size() != 0) {
+        if (!userDesignations.isEmpty()) {
             String jsmName = userDesignations.get(0).getJsm_name();
             if (jsmName == null) {
                 jsmName = "";
@@ -223,137 +222,108 @@ public class MainMenu extends AppCompatActivity {
             department.setText(deptName);
         }
 
-        personalInfo.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainMenu.this, EmplyeeInformation.class);
-                startActivity(intent);
+        personalInfo.setOnClickListener(v -> {
+            Intent intent = new Intent(MainMenu.this, EmplyeeInformation.class);
+            startActivity(intent);
 
-            }
         });
 
-        attendance.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainMenu.this, Attendance.class);
-                startActivity(intent);
-            }
+        attendance.setOnClickListener(v -> {
+            Intent intent = new Intent(MainMenu.this, Attendance.class);
+            startActivity(intent);
         });
 
-        leave.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainMenu.this, Leave.class);
-                startActivity(intent);
-            }
+        leave.setOnClickListener(v -> {
+            Intent intent = new Intent(MainMenu.this, Leave.class);
+            startActivity(intent);
         });
 
-        payRoll.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainMenu.this, PayRollInfo.class);
-                startActivity(intent);
-            }
+        payRoll.setOnClickListener(v -> {
+            Intent intent = new Intent(MainMenu.this, PayRollInfo.class);
+            startActivity(intent);
         });
 
-        directory.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainMenu.this, DirectoryWithDivision.class);
-                startActivity(intent);
-            }
+        directory.setOnClickListener(v -> {
+            Intent intent = new Intent(MainMenu.this, DirectoryWithDivision.class);
+            startActivity(intent);
         });
 
-        dashB.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainMenu.this, Dashboard.class);
-                intent.putExtra("FROMMAINMENU", false);
-                startActivity(intent);
-                finish();
-            }
+        dashB.setOnClickListener(v -> {
+            Intent intent = new Intent(MainMenu.this, Dashboard.class);
+            intent.putExtra("FROMMAINMENU", false);
+            startActivity(intent);
+            finish();
         });
 
-        logoutImage.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (isMyServiceRunning(Service.class)) {
-                    MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(MainMenu.this);
-                    builder.setMessage("Your Tracking Service is running. You can not Log Out while Running this Service!")
-                            .setPositiveButton("OK", new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialog, int which) {
-                                    dialog.dismiss();
-                                }
-                            });
-                    AlertDialog alert = builder.create();
-                    alert.show();
-                } else {
-                    MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(MainMenu.this);
-                    builder.setTitle("LOG OUT!")
-                            .setIcon(R.drawable.thrm_logo)
-                            .setMessage("Do you want to Log Out?")
-                            .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialog, int which) {
+        logoutImage.setOnClickListener(v -> {
+            if (isMyServiceRunning()) {
+                MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(MainMenu.this);
+                builder.setMessage("Your Tracking Service is running. You can not Log Out while Running this Service!")
+                        .setPositiveButton("OK", (dialog, which) -> dialog.dismiss());
+                AlertDialog alert = builder.create();
+                alert.show();
+            } else {
+                MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(MainMenu.this);
+                builder.setTitle("LOG OUT!")
+                        .setIcon(R.drawable.thrm_logo)
+                        .setMessage("Do you want to Log Out?")
+                        .setPositiveButton("YES", (dialog, which) -> {
 
 
-                                    userInfoLists.clear();
-                                    userDesignations.clear();
-                                    userInfoLists = new ArrayList<>();
-                                    userDesignations = new ArrayList<>();
-                                    isApproved = 0;
-                                    isLeaveApproved = 0;
+                            userInfoLists.clear();
+                            userDesignations.clear();
+                            userInfoLists = new ArrayList<>();
+                            userDesignations = new ArrayList<>();
+                            isApproved = 0;
+                            isLeaveApproved = 0;
 
-                                    SharedPreferences.Editor widgetEditor = attendanceWidgetPreferences.edit();
-                                    widgetEditor.remove(WIDGET_EMP_ID);
-                                    widgetEditor.remove(WIDGET_TRACKER_FLAG);
-                                    widgetEditor.apply();
-                                    widgetEditor.commit();
+                            SharedPreferences.Editor widgetEditor = attendanceWidgetPreferences.edit();
+                            widgetEditor.remove(WIDGET_EMP_ID);
+                            widgetEditor.remove(WIDGET_TRACKER_FLAG);
+                            widgetEditor.apply();
+                            widgetEditor.commit();
 
-                                    SharedPreferences.Editor editor1 = sharedPreferences.edit();
-                                    editor1.remove(USER_NAME);
-                                    editor1.remove(USER_F_NAME);
-                                    editor1.remove(USER_L_NAME);
-                                    editor1.remove(EMAIL);
-                                    editor1.remove(CONTACT);
-                                    editor1.remove(EMP_ID_LOGIN);
+                            SharedPreferences.Editor editor1 = sharedPreferences.edit();
+                            editor1.remove(USER_NAME);
+                            editor1.remove(USER_F_NAME);
+                            editor1.remove(USER_L_NAME);
+                            editor1.remove(EMAIL);
+                            editor1.remove(CONTACT);
+                            editor1.remove(EMP_ID_LOGIN);
 
-                                    editor1.remove(JSM_CODE);
-                                    editor1.remove(JSM_NAME);
-                                    editor1.remove(JSD_ID_LOGIN);
-                                    editor1.remove(JSD_OBJECTIVE);
-                                    editor1.remove(DEPT_NAME);
-                                    editor1.remove(DIV_NAME);
-                                    editor1.remove(DESG_NAME);
-                                    editor1.remove(DESG_PRIORITY);
-                                    editor1.remove(JOINING_DATE);
-                                    editor1.remove(DIV_ID);
-                                    editor1.remove(LOGIN_TF);
+                            editor1.remove(JSM_CODE);
+                            editor1.remove(JSM_NAME);
+                            editor1.remove(JSD_ID_LOGIN);
+                            editor1.remove(JSD_OBJECTIVE);
+                            editor1.remove(DEPT_NAME);
+                            editor1.remove(DIV_NAME);
+                            editor1.remove(DESG_NAME);
+                            editor1.remove(DESG_PRIORITY);
+                            editor1.remove(JOINING_DATE);
+                            editor1.remove(DIV_ID);
+                            editor1.remove(LOGIN_TF);
 
-                                    editor1.remove(IS_ATT_APPROVED);
-                                    editor1.remove(IS_LEAVE_APPROVED);
-                                    editor1.remove(COMPANY);
-                                    editor1.remove(SOFTWARE);
-                                    editor1.remove(LIVE_FLAG);
+                            editor1.remove(IS_ATT_APPROVED);
+                            editor1.remove(IS_LEAVE_APPROVED);
+                            editor1.remove(COMPANY);
+                            editor1.remove(SOFTWARE);
+                            editor1.remove(LIVE_FLAG);
 //                                    editor1.remove(DATABASE_NAME);
-                                    editor1.apply();
-                                    editor1.commit();
+                            editor1.apply();
+                            editor1.commit();
 
-                                    if (trackerAvailable == 1) {
-                                        SharedPreferences.Editor editor = sharedSchedule.edit();
-                                        editor.remove(SCHEDULING_EMP_ID);
-                                        editor.remove(TRIGGERING);
-                                        editor.apply();
-                                        editor.commit();
+                            if (trackerAvailable == 1) {
+                                SharedPreferences.Editor editor = sharedSchedule.edit();
+                                editor.remove(SCHEDULING_EMP_ID);
+                                editor.remove(TRIGGERING);
+                                editor.apply();
+                                editor.commit();
 
-                                        Intent intent1 = new Intent(MainMenu.this, Uploader.class);
-                                        PendingIntent pendingIntent = null;
-                                        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
-                                            pendingIntent = PendingIntent.getBroadcast(getApplicationContext(), 0, intent1,PendingIntent.FLAG_IMMUTABLE);
-                                        }
-                                        alarmManager.cancel(pendingIntent);
-                                    }
+                                Intent intent1 = new Intent(MainMenu.this, Uploader.class);
+                                PendingIntent pendingIntent;
+                                pendingIntent = PendingIntent.getBroadcast(getApplicationContext(), 0, intent1, PendingIntent.FLAG_IMMUTABLE);
+                                alarmManager.cancel(pendingIntent);
+                            }
 
 //                        if (Build.VERSION.SDK_INT < 16) {
 //
@@ -366,93 +336,44 @@ public class MainMenu extends AppCompatActivity {
 //                            int uiOptions = View.SYSTEM_UI_FLAG_FULLSCREEN;
 //                            decorView.setSystemUiVisibility(uiOptions);
 //                        }
-                                    Intent intent = new Intent(MainMenu.this, Login.class);
-                                    startActivity(intent);
-                                    finish();
-                                    //System.exit(0);
-                                }
-                            })
-                            .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialog, int which) {
+                            Intent intent = new Intent(MainMenu.this, Login.class);
+                            startActivity(intent);
+                            finish();
+                            //System.exit(0);
+                        })
+                        .setNegativeButton("NO", (dialog, which) -> {
 
-                                }
-                            });
-                    AlertDialog alert = builder.create();
-                    alert.show();
-                }
+                        });
+                AlertDialog alert = builder.create();
+                alert.show();
+            }
 
+        });
+
+        getOnBackPressedDispatcher().addCallback(this, new OnBackPressedCallback(true) {
+            @Override
+            public void handleOnBackPressed() {
+                MaterialAlertDialogBuilder alertDialogBuilder = new MaterialAlertDialogBuilder(MainMenu.this)
+                        .setTitle("EXIT!")
+                        .setMessage("Do You Want to Exit?")
+                        .setIcon(R.drawable.thrm_logo)
+                        .setPositiveButton("Yes", (dialog, which) -> finish())
+                        .setNegativeButton("No", (dialog, which) -> {
+                            //Do nothing
+                        });
+                AlertDialog alertDialog = alertDialogBuilder.create();
+                alertDialog.show();
             }
         });
     }
 
-    private boolean isMyServiceRunning(Class<?> serviceClass) {
+    private boolean isMyServiceRunning() {
         ActivityManager manager = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
         for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
-            if (serviceClass.getName().equals(service.service.getClassName())) {
+            if (Service.class.getName().equals(service.service.getClassName())) {
                 return true;
             }
         }
         return false;
-    }
-
-    @Override
-    public void onBackPressed() {
-//        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-//        builder.setTitle("LOG OUT!")
-//                .setMessage("Do you want to Log Out?")
-//                .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-//                    @Override
-//                    public void onClick(DialogInterface dialog, int which) {
-//
-//
-//                        userInfoLists.clear();
-//                        userDesignations.clear();
-//                        userInfoLists = new ArrayList<>();
-//                        userDesignations = new ArrayList<>();
-//                        isApproved = 0;
-//                        isLeaveApproved = 0;
-////                        if (Build.VERSION.SDK_INT < 16) {
-////
-////                            window.setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-////                                    WindowManager.LayoutParams.FLAG_FULLSCREEN);
-////                        }
-////                        else {
-////
-////                            // Hide Status Bar.
-////                            int uiOptions = View.SYSTEM_UI_FLAG_FULLSCREEN;
-////                            decorView.setSystemUiVisibility(uiOptions);
-////                        }
-//                        finish();
-//                        //System.exit(0);
-//                    }
-//                })
-//                .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-//                    @Override
-//                    public void onClick(DialogInterface dialog, int which) {
-//
-//                    }
-//                });
-//        AlertDialog alert = builder.create();
-//        alert.show();
-        MaterialAlertDialogBuilder alertDialogBuilder = new MaterialAlertDialogBuilder(MainMenu.this)
-                .setTitle("EXIT!")
-                .setMessage("Do You Want to Exit?")
-                .setIcon(R.drawable.thrm_logo)
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        finish();
-                    }
-                })
-                .setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        //Do nothing
-                    }
-                });
-        AlertDialog alertDialog = alertDialogBuilder.create();
-        alertDialog.show();
-
     }
 }
