@@ -1,7 +1,6 @@
 package ttit.com.shuvo.ikglhrm.attendance.giveAttendance;
 
-import androidx.activity.result.ActivityResult;
-import androidx.activity.result.ActivityResultCallback;
+import androidx.activity.OnBackPressedCallback;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
@@ -13,12 +12,10 @@ import androidx.core.app.ActivityCompat;
 import androidx.core.content.res.ResourcesCompat;
 
 import android.Manifest;
-import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.ActivityManager;
 import android.content.ComponentName;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentSender;
 import android.content.SharedPreferences;
@@ -28,9 +25,6 @@ import android.location.Address;
 import android.location.Geocoder;
 import android.location.Location;
 import android.location.LocationManager;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Looper;
 import android.os.StrictMode;
@@ -42,66 +36,63 @@ import android.widget.TextClock;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.android.volley.AuthFailureError;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.google.android.gms.common.ConnectionResult;
-import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.common.api.PendingResult;
-import com.google.android.gms.common.api.ResultCallback;
-import com.google.android.gms.common.api.Status;
+import com.google.android.gms.common.api.ResolvableApiException;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationCallback;
 import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationResult;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.location.LocationSettingsRequest;
-import com.google.android.gms.location.LocationSettingsResult;
-import com.google.android.gms.location.LocationSettingsStates;
-import com.google.android.gms.location.LocationSettingsStatusCodes;
+import com.google.android.gms.location.LocationSettingsResponse;
+import com.google.android.gms.location.Priority;
+import com.google.android.gms.location.SettingsClient;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.CircleOptions;
 import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 
 
 import java.io.IOException;
 import java.sql.Timestamp;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import ttit.com.shuvo.ikglhrm.R;
 import ttit.com.shuvo.ikglhrm.WaitProgress;
+import ttit.com.shuvo.ikglhrm.attendance.giveAttendance.arraylists.AreaList;
 import ttit.com.shuvo.ikglhrm.attendance.trackService.Service;
 
-import static ttit.com.shuvo.ikglhrm.Login.SoftwareName;
 import static ttit.com.shuvo.ikglhrm.Login.userInfoLists;
 import static ttit.com.shuvo.ikglhrm.attendance.Attendance.live_tracking_flag;
 import static ttit.com.shuvo.ikglhrm.attendance.Attendance.tracking_flag;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallback, GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
+public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallback {
 
     private GoogleMap mMap;
     TextView currLoc;
     TextView checkInTime;
     CardView chekInButton;
     TextView nameOfCheckIN;
-    Button close;
-    private GoogleApiClient googleApiClient;
     FusedLocationProviderClient fusedLocationProviderClient;
     LocationManager locationManager;
     LocationRequest locationRequest;
@@ -116,40 +107,54 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
     SharedPreferences preferences;
 
     WaitProgress waitProgress = new WaitProgress();
-    private String message = null;
     private Boolean conn = false;
     private Boolean connected = false;
 //    private Connection connection;
     Timestamp ts;
 
-    TextView software;
+//    TextView software;
     ImageView autoStartIcon;
 
     ActivityResultLauncher<Intent> someActivityResultLauncher;
-    String officeLatitude = "";
-    String officeLongitude = "";
-    String coverage = "";
-    String machineCode = "3";
+
+    String machineCode = "";
     String last_time = "";
     String today_date = "";
     String timeToShow = "";
     TextClock digitalClock;
     TextView todayTime;
 
+    ArrayList<AreaList> areaLists;
+
+    Logger logger = Logger.getLogger(AttendanceGive.class.getName());
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_attendance_give);
+
+        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
+                .findFragmentById(R.id.map);
+        assert mapFragment != null;
+        mapFragment.getMapAsync(this);
+
         currLoc = findViewById(R.id.text_of_cu_loc);
         currLoc.setVisibility(View.GONE);
         checkInTime = findViewById(R.id.check_int_time);
         chekInButton = findViewById(R.id.check_in_time_button);
         nameOfCheckIN = findViewById(R.id.name_of_punch);
-        close = findViewById(R.id.give_att_finish);
-        software = findViewById(R.id.name_of_company_attendance_give);
+//        software = findViewById(R.id.name_of_company_attendance_give);
         autoStartIcon = findViewById(R.id.app_auto_start_icon);
         digitalClock = findViewById(R.id.text_clock_give_att);
         todayTime = findViewById(R.id.today_date_time_give_att);
+
+        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(this);
+        locationManager = (LocationManager) this.getSystemService(Context.LOCATION_SERVICE);
+        locationRequest = new LocationRequest.Builder(Priority.PRIORITY_HIGH_ACCURACY, 5000)
+                .setWaitForAccurateLocation(false)
+                .setMinUpdateIntervalMillis(1000)
+                .setMaxUpdateDelayMillis(2000)
+                .build();
 
         Intent intent = getIntent();
         last_time = intent.getStringExtra("LAST_TIME");
@@ -160,40 +165,29 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
 
         todayTime.setText(today_date);
 
-        software.setText(SoftwareName);
+//        software.setText(SoftwareName);
 
         if (tracking_flag == 1) {
 
-            if (isMyServiceRunning(Service.class)) {
-                nameOfCheckIN.setText("PUNCH & STOP TRACKER");
+            if (isMyServiceRunning()) {
+                String tt = "PUNCH & STOP TRACKER";
+                nameOfCheckIN.setText(tt);
             } else {
-                nameOfCheckIN.setText("PUNCH & START TRACKER");
+                String tt = "PUNCH & START TRACKER";
+                nameOfCheckIN.setText(tt);
             }
 
         } else {
-            nameOfCheckIN.setText("PUNCH");
+            String  tt = "PUNCH";
+            nameOfCheckIN.setText(tt);
         }
-
-        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(this);
-
-        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
-                .findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-        locationManager = (LocationManager) this.getSystemService(Context.LOCATION_SERVICE);
-        locationRequest = LocationRequest.create();
-        locationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
-//        locationRequest.setInterval(5);
-        locationRequest.setInterval(5000);
-        locationRequest.setFastestInterval(1000);
 
         emp_id = userInfoLists.get(0).getEmp_id();
 
         preferences = getSharedPreferences(emp_id,MODE_PRIVATE);
 
         getTime = preferences.getString(timeKey,null);
-//        if (getTime != null) {
-//            checkInTime.setText(getTime);
-//        }
+
         String lt = "Your last recorded time : " + last_time;
         checkInTime.setText(lt);
 
@@ -202,23 +196,16 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
 
         someActivityResultLauncher = registerForActivityResult(
                 new ActivityResultContracts.StartActivityForResult(),
-                new ActivityResultCallback<ActivityResult>() {
-                    @Override
-                    public void onActivityResult(ActivityResult result) {
-                        if (result.getResultCode() == Activity.RESULT_OK) {
-                            // There are no request codes
-                            Intent data = result.getData();
-                            System.out.println("EKHANE ASHE CHECK: " + data.toString());
-                        }
+                result -> {
+                    if (result.getResultCode() == Activity.RESULT_OK) {
+                        // There are no request codes
+                        Intent data = result.getData();
+                        assert data != null;
+                        System.out.println("EKHANE ASHE CHECK: " + data);
                     }
                 });
 
-        autoStartIcon.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                PermissionsAll();
-            }
-        });
+        autoStartIcon.setOnClickListener(v -> PermissionsAll());
 
     }
 
@@ -247,9 +234,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setTitle("Check Auto Start Permission!")
                 .setMessage("Check the App Auto Start Option is On or Off. Auto Start On will provide better solution for the service in the background.")
-                .setPositiveButton("Check", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
+                .setPositiveButton("Check", (dialog, which) -> {
 
 //                        for (Intent intent : POWERMANAGER_INTENTS)
 //                            if (getPackageManager().resolveActivity(intent, PackageManager.MATCH_DEFAULT_ONLY) != null) {
@@ -260,23 +245,19 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
 //                                break;
 //                            }
 
-                        for (Intent intent : POWERMANAGER_INTENTS)
-                            if (getPackageManager().resolveActivity(intent, PackageManager.MATCH_DEFAULT_ONLY) != null) {
-                                System.out.println("PAISE KISU MISU: "+ intent.getComponent().toString());
-                                paise[0] = true;
-                                someActivityResultLauncher.launch(intent);
-                                break;
-                            }
-                        if (!paise[0]){
-                            Toast.makeText(getApplicationContext(),"Could not find Auto Start Permission Settings.",Toast.LENGTH_SHORT).show();
+                    for (Intent intent : POWERMANAGER_INTENTS)
+                        if (getPackageManager().resolveActivity(intent, PackageManager.MATCH_DEFAULT_ONLY) != null) {
+                            System.out.println("PAISE KISU MISU: "+ intent.getComponent().toString());
+                            paise[0] = true;
+                            someActivityResultLauncher.launch(intent);
+                            break;
                         }
+                    if (!paise[0]){
+                        Toast.makeText(getApplicationContext(),"Could not find Auto Start Permission Settings.",Toast.LENGTH_SHORT).show();
                     }
                 })
-                .setNegativeButton("Don't Check", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
+                .setNegativeButton("Don't Check", (dialog, which) -> {
 
-                    }
                 });
         AlertDialog alert = builder.create();
         alert.show();
@@ -300,7 +281,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
     }
 
     @Override
-    public void onMapReady(GoogleMap googleMap) {
+    public void onMapReady(@NonNull GoogleMap googleMap) {
         mMap = googleMap;
         final LatLng[] lastLatLongitude = {new LatLng(0, 0)};
 
@@ -311,9 +292,6 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
         locationCallback = new LocationCallback() {
             @Override
             public void onLocationResult(@NonNull LocationResult locationResult) {
-                if (locationResult == null) {
-                    return;
-                }
                 for (Location location : locationResult.getLocations()) {
 
                     Log.i("LocationFused ", location.toString());
@@ -325,7 +303,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                     Date c = Calendar.getInstance().getTime();
                     Date date = new Date();
                     ts = new Timestamp(date.getTime());
-                    System.out.println(ts.toString());
+                    System.out.println(ts);
                     inTime = df.format(c);
                     timeToShow = dftoShow.format(c);
                     System.out.println("IN TIME : " + inTime);
@@ -335,201 +313,192 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
             }
         };
 
-        chekInButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (!inTime.isEmpty()) {
-                    LatLng c_latLng = new LatLng(0,0);
-                    if (officeLatitude != null && officeLongitude != null) {
-                        if (!officeLatitude.isEmpty() && !officeLongitude.isEmpty()) {
-                            c_latLng = new LatLng(Double.parseDouble(officeLatitude),Double.parseDouble(officeLongitude));
-                        }
-                    }
-                    if (c_latLng.latitude != 0 && c_latLng.longitude != 0) {
+        chekInButton.setOnClickListener(v -> {
+            if (!inTime.isEmpty()) {
+                if (!areaLists.isEmpty()) {
+                    if (tracking_flag == 1) {
+                        LatLng c_latLng = new LatLng(0,0);
                         float[] distance = new float[1];
-                        Location.distanceBetween(c_latLng.latitude,c_latLng.longitude,lastLatLongitude[0].latitude,lastLatLongitude[0].longitude,distance);
+                        float prev_distance = 0;
+                        String prev_mach_code = "";
 
-                        float radius = 0;
-                        if (coverage != null) {
-                            if (!coverage.isEmpty()) {
-                                radius = Float.parseFloat(coverage);
-                            }
-                        }
+                        machineCode = "";
 
-                        if (radius == 0) {
-                            machineCode = "3";
-                        } else {
-                            machineCode = "1";
-                        }
+                        for (int i = 0; i < areaLists.size(); i++) {
+                            String officeLatitude = areaLists.get(i).getLatitude();
+                            String officeLongitude = areaLists.get(i).getLongitude();
+                            String coverage = areaLists.get(i).getCoverage();
 
-                        if (distance[0] <= radius || radius == 0) {
-                            if (tracking_flag == 1 ) {
-                                if (isMyServiceRunning(Service.class)) {
-
-                                    AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
-                                    builder.setTitle("Attendance!")
-                                            .setMessage("Do you want to punch & stop your tracker?")
-                                            .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog, int which) {
-
-
-                                                    new CheckAddress().execute();
-
-                                                }
-                                            })
-                                            .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog, int which) {
-
-                                                }
-                                            });
-                                    AlertDialog alert = builder.create();
-                                    alert.show();
-                                } else {
-                                    AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
-                                    builder.setTitle("Attendance!")
-                                            .setMessage("Do you want to punch & start your tracker?")
-                                            .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog, int which) {
-
-
-                                                    new CheckAddress().execute();
-
-                                                }
-                                            })
-                                            .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                                @Override
-                                                public void onClick(DialogInterface dialog, int which) {
-
-                                                }
-                                            });
-                                    AlertDialog alert = builder.create();
-                                    alert.show();
+                            if (officeLatitude != null && officeLongitude != null) {
+                                if (!officeLatitude.isEmpty() && !officeLongitude.isEmpty()) {
+                                    c_latLng = new LatLng(Double.parseDouble(officeLatitude),Double.parseDouble(officeLongitude));
                                 }
-                            } else {
-                                AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
-                                builder.setTitle("Punch Attendance!")
-                                        .setMessage("Do you want to punch now?")
-                                        .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-
-
-                                                new CheckAddress().execute();
-
-                                            }
-                                        })
-                                        .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-
-                                            }
-                                        });
-                                AlertDialog alert = builder.create();
-                                alert.show();
                             }
+
+                            if (c_latLng.latitude != 0 && c_latLng.longitude != 0) {
+                                Location.distanceBetween(c_latLng.latitude,c_latLng.longitude,lastLatLongitude[0].latitude,lastLatLongitude[0].longitude,distance);
+
+                                float radius = 0;
+                                if (coverage != null) {
+                                    if (!coverage.isEmpty()) {
+                                        radius = Float.parseFloat(coverage);
+                                    }
+                                }
+
+                                machineCode = areaLists.get(i).getMachine_code();
+
+                                if (distance[0] <= radius) {
+                                    prev_mach_code = machineCode;
+                                    break;
+                                }
+                                else {
+                                    float dd = distance[0] - radius;
+                                    if (prev_distance == 0) {
+                                        prev_distance = dd;
+                                        prev_mach_code = machineCode;
+                                    }
+                                    else if (dd < prev_distance) {
+                                        prev_distance = dd;
+                                        prev_mach_code = machineCode;
+                                    }
+                                }
+                            }
+                        }
+
+                        machineCode = prev_mach_code;
+
+                        if (machineCode.isEmpty()) {
+                            machineCode = areaLists.get(0).getMachine_code();
+                        }
+
+                        if (isMyServiceRunning()) {
+                            AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
+                            builder.setTitle("Attendance!")
+                                    .setMessage("Do you want to punch & stop your tracker?")
+                                    .setPositiveButton("YES", (dialog, which) -> checkAddress())
+                                    .setNegativeButton("NO", (dialog, which) -> {
+                                    });
+                            AlertDialog alert = builder.create();
+                            alert.show();
                         }
                         else {
-                            Toast.makeText(getApplicationContext(),"You are not around your office area",Toast.LENGTH_SHORT).show();
-                        }
-                    }
-                    else {
-                        machineCode = "3";
-                        if (tracking_flag == 1) {
-                            if (isMyServiceRunning(Service.class)) {
-
-                                AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
-                                builder.setTitle("Attendance!")
-                                        .setMessage("Do you want to punch & stop your tracker?")
-                                        .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-
-
-                                                new CheckAddress().execute();
-
-                                            }
-                                        })
-                                        .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-
-                                            }
-                                        });
-                                AlertDialog alert = builder.create();
-                                alert.show();
-                            } else {
-                                AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
-                                builder.setTitle("Attendance!")
-                                        .setMessage("Do you want to punch & start your tracker?")
-                                        .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-
-
-                                                new CheckAddress().execute();
-
-                                            }
-                                        })
-                                        .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-
-                                            }
-                                        });
-                                AlertDialog alert = builder.create();
-                                alert.show();
-                            }
-                        } else {
                             AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
-                            builder.setTitle("Punch Attendance!")
-                                    .setMessage("Do you want to punch now?")
-                                    .setPositiveButton("YES", new DialogInterface.OnClickListener() {
-                                        @Override
-                                        public void onClick(DialogInterface dialog, int which) {
-
-
-                                            new CheckAddress().execute();
-
-                                        }
-                                    })
-                                    .setNegativeButton("NO", new DialogInterface.OnClickListener() {
-                                        @Override
-                                        public void onClick(DialogInterface dialog, int which) {
-
-                                        }
+                            builder.setTitle("Attendance!")
+                                    .setMessage("Do you want to punch & start your tracker?")
+                                    .setPositiveButton("YES", (dialog, which) -> checkAddress())
+                                    .setNegativeButton("NO", (dialog, which) -> {
                                     });
                             AlertDialog alert = builder.create();
                             alert.show();
                         }
                     }
+                    else {
+                        LatLng c_latLng = new LatLng(0,0);
+                        float[] distance = new float[1];
+                        boolean found = false;
+                        float prev_distance = 0;
+                        String prev_mach_code = "";
 
-                } else {
-                    Toast.makeText(getApplicationContext(),"Please wait for getting the location",Toast.LENGTH_SHORT).show();
+                        machineCode = "";
+
+                        for (int i = 0; i < areaLists.size(); i++) {
+                            String officeLatitude = areaLists.get(i).getLatitude();
+                            String officeLongitude = areaLists.get(i).getLongitude();
+                            String coverage = areaLists.get(i).getCoverage();
+
+                            if (officeLatitude != null && officeLongitude != null) {
+                                if (!officeLatitude.isEmpty() && !officeLongitude.isEmpty()) {
+                                    c_latLng = new LatLng(Double.parseDouble(officeLatitude),Double.parseDouble(officeLongitude));
+                                }
+                            }
+
+                            if (c_latLng.latitude != 0 && c_latLng.longitude != 0) {
+                                Location.distanceBetween(c_latLng.latitude,c_latLng.longitude,lastLatLongitude[0].latitude,lastLatLongitude[0].longitude,distance);
+
+                                float radius = 0;
+                                if (coverage != null) {
+                                    if (!coverage.isEmpty()) {
+                                        radius = Float.parseFloat(coverage);
+                                    }
+                                }
+
+                                machineCode = areaLists.get(i).getMachine_code();
+
+                                if (distance[0] <= radius) {
+                                    found = true;
+                                    prev_mach_code = machineCode;
+                                    break;
+                                }
+                                else {
+                                    float dd = distance[0] - radius;
+                                    if (prev_distance == 0) {
+                                        prev_distance = dd;
+                                        prev_mach_code = machineCode;
+                                    }
+                                    else if (dd < prev_distance) {
+                                        prev_distance = dd;
+                                        prev_mach_code = machineCode;
+                                    }
+                                }
+                            }
+                        }
+
+                        machineCode = prev_mach_code;
+
+                        if (found) {
+                            AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
+                            builder.setTitle("Punch Attendance!")
+                                    .setMessage("Do you want to punch now?")
+                                    .setPositiveButton("YES", (dialog, which) -> checkAddress())
+                                    .setNegativeButton("NO", (dialog, which) -> {
+                                    });
+                            AlertDialog alert = builder.create();
+                            alert.show();
+                        }
+                        else {
+                            if (areaLists.get(0).isCanGive()) {
+                                if (machineCode.isEmpty()) {
+                                    machineCode = areaLists.get(0).getMachine_code();
+                                }
+                                AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
+                                builder.setTitle("Punch Attendance!")
+                                        .setMessage("Do you want to punch now?")
+                                        .setPositiveButton("YES", (dialog, which) -> checkAddress())
+                                        .setNegativeButton("NO", (dialog, which) -> {
+                                        });
+                                AlertDialog alert = builder.create();
+                                alert.show();
+                            }
+                            else {
+                                if (prev_distance == 0) {
+                                    Toast.makeText(getApplicationContext(),"You are not around your office area",Toast.LENGTH_SHORT).show();
+                                }
+                                else {
+                                    Toast.makeText(getApplicationContext(),"You are not around your office area. You are "+prev_distance+" Meters away from office.",Toast.LENGTH_LONG).show();
+                                }
+                            }
+                        }
+                    }
                 }
+                else {
+                    Toast.makeText(getApplicationContext(),"You don't have any permission to give attendance from this app. Please contact with administrator",Toast.LENGTH_LONG).show();
+                }
+            }
+            else {
+                Toast.makeText(getApplicationContext(),"Please wait for getting the location",Toast.LENGTH_SHORT).show();
             }
         });
 
-        close.setOnClickListener(new View.OnClickListener() {
+        getOnBackPressedDispatcher().addCallback(this, new OnBackPressedCallback(true) {
             @Override
-            public void onClick(View v) {
+            public void handleOnBackPressed() {
                 fusedLocationProviderClient.removeLocationUpdates(locationCallback);
                 finish();
             }
         });
 
-//        new LocationGet().execute();
         getOfficeLocation();
     }
-
-    @Override
-    public void onBackPressed() {
-        fusedLocationProviderClient.removeLocationUpdates(locationCallback);
-        finish();
-    }
-
 
     public void getAddress(double lat, double lng) {
         Geocoder geocoder = new Geocoder(AttendanceGive.this, Locale.ENGLISH);
@@ -537,149 +506,153 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
             List<Address> addresses = geocoder.getFromLocation(lat, lng, 1);
             if (Geocoder.isPresent()) {
                 Address obj = addresses.get(0);
-                String adds = obj.getAddressLine(0);
-//                String add = "Address from GeoCODE: ";
-//                add = add + "\n" + obj.getCountryName();
-//                add = add + "\n" + obj.getCountryCode();
-//                add = add + "\n" + obj.getAdminArea();
-//                add = add + "\n" + obj.getPostalCode();
-//                add = add + "\n" + obj.getSubAdminArea();
-//                add = add + "\n" + obj.getLocality();
-//                add = add + "\n" + obj.getSubThoroughfare();
-//                add = add + "\n" + obj.getFeatureName();
-//                add = add + "\n" + obj.getPhone();
-//                add = add + "\n" + obj.getPremises();
-//                add = add + "\n" + obj.getSubLocality();
-//                add = add + "\n" + obj.getThoroughfare();
-//                add = add + "\n" + obj.getUrl();
-
-//                Log.v("IGA", "Address: " + add);
-//                Log.v("NEW ADD", "Address: " + adds);
-                address = adds;
+                address = obj.getAddressLine(0);
                 System.out.println("Ekhane ashbe 1st");
-                // Toast.makeText(this, "Address=>" + add,
-                // Toast.LENGTH_SHORT).show();
-
             } else {
                 address = "";
             }
-
-            // TennisAppActivity.showDialog(add);
-        } catch (IOException e) {
-            // TODO Auto-generated catch block
-            e.printStackTrace();
+        }
+        catch (IOException e) {
             address = "";
-            //Toast.makeText(this, e.getMessage(), Toast.LENGTH_SHORT).show();
         }
     }
 
-    @SuppressLint("MissingPermission")
     public void zoomToUserLocation(){
-
         if (ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            // TODO: Consider calling
             Log.i("Ekhane", "1");
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
             return;
         }
         mMap.setMyLocationEnabled(true);
         Task<Location> locationTask = fusedLocationProviderClient.getLastLocation();
-        locationTask.addOnSuccessListener(new OnSuccessListener<Location>() {
-            @Override
-            public void onSuccess(Location location) {
-//                Log.i("lattt", location.toString());
-                LatLng latLng = new LatLng(23.6850, 90.3563);
-
-
-                if (location != null) {
-                    latLng = new LatLng(location.getLatitude(), location.getLongitude());
-                    mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(latLng, 18));
-                } else {
-                    latLng = new LatLng(23.6850, 90.3563);
-                    mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(latLng, 7));
-                }
-
+        locationTask.addOnSuccessListener(location -> {
+            LatLng latLng;
+            if (location != null) {
+                latLng = new LatLng(location.getLatitude(), location.getLongitude());
+                mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(latLng, 18));
+            } else {
+                latLng = new LatLng(23.6850, 90.3563);
+                mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(latLng, 7));
             }
+
         });
 
         fusedLocationProviderClient.requestLocationUpdates(locationRequest, locationCallback, Looper.getMainLooper());
 
-
-
-    }
-    private void enableGPS() {
-        if (googleApiClient == null) {
-            googleApiClient = new GoogleApiClient.Builder(this)
-                    .addApi(LocationServices.API).addConnectionCallbacks(AttendanceGive.this)
-                    .addOnConnectionFailedListener(AttendanceGive.this).build();
-            googleApiClient.connect();
-            LocationRequest locationRequest = LocationRequest.create();
-            locationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
-            locationRequest.setInterval(5 * 1000);
-            locationRequest.setFastestInterval(1000);
-            LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder()
-                    .addLocationRequest(locationRequest);
-
-            // **************************
-            builder.setAlwaysShow(true); // this is the key ingredient
-            // **************************
-
-            PendingResult<LocationSettingsResult> result = LocationServices.SettingsApi
-                    .checkLocationSettings(googleApiClient, builder.build());
-            result.setResultCallback(new ResultCallback<LocationSettingsResult>() {
-                @Override
-                public void onResult(LocationSettingsResult result) {
-                    final Status status = result.getStatus();
-                    final LocationSettingsStates state = result
-                            .getLocationSettingsStates();
-                    switch (status.getStatusCode()) {
-                        case LocationSettingsStatusCodes.SUCCESS:
-                            // All location settings are satisfied. The client can
-                            // initialize location
-                            // requests here.
-                            Log.i("Exit", "3");
-                            zoomToUserLocation();
-                            //info.setText("Done");
-
-                            break;
-                        case LocationSettingsStatusCodes.RESOLUTION_REQUIRED:
-                            // Location settings are not satisfied. But could be
-                            // fixed by showing the user
-                            // a dialog.
-                            Log.i("Exit", "4");
-                            try {
-                                // Show the dialog by calling
-                                // startResolutionForResult(),
-                                // and check the result in onActivityResult().
-                                status.startResolutionForResult(AttendanceGive.this, 1000);
-                            } catch (IntentSender.SendIntentException e) {
-                                // Ignore the error.
-                            }
-                            break;
-                        case LocationSettingsStatusCodes.SETTINGS_CHANGE_UNAVAILABLE:
-                            // Location settings are not satisfied. However, we have
-                            // no way to fix the
-                            // settings so we won't show the dialog.
-                            Log.i("Exit", "5");
-                            break;
-                    }
+        for (int i = 0; i < areaLists.size(); i++) {
+            if (areaLists.get(i).getLatitude() != null && areaLists.get(i).getLongitude() != null && areaLists.get(i).getCoverage() != null)  {
+                if (!areaLists.get(i).getCoverage().equals("0")) {
+                    mMap.addCircle(new CircleOptions()
+                            .center(new LatLng(Float.parseFloat(areaLists.get(i).getLatitude()), Float.parseFloat(areaLists.get(i).getLongitude())))
+                            .radius(Integer.parseInt(areaLists.get(i).getCoverage()))
+                            .strokeColor(getColor(R.color.primaryColor))
+                            .strokeWidth(4F)
+                            .fillColor(getColor(R.color.primaryColor_generatedAlpha)));
                 }
-
-            });
-
-
+            }
         }
+
+
     }
 
-    private boolean isMyServiceRunning(Class<?> serviceClass) {
+//    private void enableGPS() {
+//        if (googleApiClient == null) {
+//            googleApiClient = new GoogleApiClient.Builder(this)
+//                    .addApi(LocationServices.API).addConnectionCallbacks(AttendanceGive.this)
+//                    .addOnConnectionFailedListener(AttendanceGive.this).build();
+//            googleApiClient.connect();
+//            LocationRequest locationRequest = LocationRequest.create();
+//            locationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
+//            locationRequest.setInterval(5 * 1000);
+//            locationRequest.setFastestInterval(1000);
+//            LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder()
+//                    .addLocationRequest(locationRequest);
+//
+//            // **************************
+//            builder.setAlwaysShow(true); // this is the key ingredient
+//            // **************************
+//
+//            PendingResult<LocationSettingsResult> result = LocationServices.SettingsApi
+//                    .checkLocationSettings(googleApiClient, builder.build());
+//            result.setResultCallback(new ResultCallback<LocationSettingsResult>() {
+//                @Override
+//                public void onResult(LocationSettingsResult result) {
+//                    final Status status = result.getStatus();
+//                    final LocationSettingsStates state = result
+//                            .getLocationSettingsStates();
+//                    switch (status.getStatusCode()) {
+//                        case LocationSettingsStatusCodes.SUCCESS:
+//                            // All location settings are satisfied. The client can
+//                            // initialize location
+//                            // requests here.
+//                            Log.i("Exit", "3");
+//                            zoomToUserLocation();
+//                            //info.setText("Done");
+//
+//                            break;
+//                        case LocationSettingsStatusCodes.RESOLUTION_REQUIRED:
+//                            // Location settings are not satisfied. But could be
+//                            // fixed by showing the user
+//                            // a dialog.
+//                            Log.i("Exit", "4");
+//                            try {
+//                                // Show the dialog by calling
+//                                // startResolutionForResult(),
+//                                // and check the result in onActivityResult().
+//                                status.startResolutionForResult(AttendanceGive.this, 1000);
+//                            } catch (IntentSender.SendIntentException e) {
+//                                // Ignore the error.
+//                            }
+//                            break;
+//                        case LocationSettingsStatusCodes.SETTINGS_CHANGE_UNAVAILABLE:
+//                            // Location settings are not satisfied. However, we have
+//                            // no way to fix the
+//                            // settings so we won't show the dialog.
+//                            Log.i("Exit", "5");
+//                            break;
+//                    }
+//                }
+//
+//            });
+//
+//
+//        }
+//    }
+
+    private void enableGPS() {
+        LocationRequest locationRequest = new LocationRequest.Builder(Priority.PRIORITY_HIGH_ACCURACY, 3000)
+                .setWaitForAccurateLocation(false)
+                .setMinUpdateIntervalMillis(1000)
+                .setMaxUpdateDelayMillis(2000)
+                .build();
+        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder()
+                .addLocationRequest(locationRequest);
+
+        SettingsClient client = LocationServices.getSettingsClient(this);
+        Task<LocationSettingsResponse> task = client.checkLocationSettings(builder.build());
+
+        task.addOnSuccessListener(this, locationSettingsResponse -> zoomToUserLocation());
+
+        task.addOnFailureListener(this, e -> {
+            if (e instanceof ResolvableApiException) {
+                // Location settings are not satisfied, but this can be fixed
+                // by showing the user a dialog.
+                try {
+                    // Show the dialog by calling startResolutionForResult(),
+                    // and check the result in onActivityResult().
+                    ResolvableApiException resolvable = (ResolvableApiException) e;
+                    resolvable.startResolutionForResult(AttendanceGive.this,
+                            1000);
+                } catch (IntentSender.SendIntentException sendEx) {
+                    // Ignore the error.
+                }
+            }
+        });
+    }
+
+    private boolean isMyServiceRunning() {
         ActivityManager manager = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
         for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
-            if (serviceClass.getName().equals(service.service.getClassName())) {
+            if (Service.class.getName().equals(service.service.getClassName())) {
                 return true;
             }
         }
@@ -691,125 +664,138 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
         super.onActivityResult(requestCode, resultCode, data);
         if (requestCode == 1000) {
             if (resultCode == Activity.RESULT_OK) {
-                String result = data.getStringExtra("result");
-                //info.setText("Done");
                 zoomToUserLocation();
                 Log.i("Hoise ", "1");
             }
             if (resultCode == Activity.RESULT_CANCELED) {
-                //Write your code if there's no result
                 Log.i("Hoise ", "2");
                 finish();
-//                System.exit(0);
             }
         }
     }
 
-    @Override
-    public void onConnected(@Nullable Bundle bundle) {
+    public void checkAddress() {
+        waitProgress.show(getSupportFragmentManager(),"WaitBar");
+        waitProgress.setCancelable(false);
+        new Thread(() -> {
+            getAddress(Double.parseDouble(lat),Double.parseDouble(lon));
+            runOnUiThread(() -> {
+                if (!address.isEmpty()) {
+                    giveAttendance();
+                }
+                else {
+                    waitProgress.dismiss();
+                    AlertDialog dialog = new AlertDialog.Builder(AttendanceGive.this)
+                            .setMessage("Could not get address of the location due to internet disruption. Please try again")
+                            .setPositiveButton("Retry", null)
+                            .setNegativeButton("Cancel",null)
+                            .show();
 
-    }
+                    dialog.setCancelable(false);
+                    dialog.setCanceledOnTouchOutside(false);
+                    Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
+                    positive.setOnClickListener(v -> {
 
-    @Override
-    public void onConnectionSuspended(int i) {
-
-    }
-
-    @Override
-    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
-
-    }
-
-    public boolean isConnected() {
-        boolean connected = false;
-        boolean isMobile = false;
-        try {
-            ConnectivityManager cm = (ConnectivityManager) getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
-            @SuppressLint("MissingPermission") NetworkInfo nInfo = cm.getActiveNetworkInfo();
-            connected = nInfo != null && nInfo.isAvailable() && nInfo.isConnected();
-            return connected;
-        } catch (Exception e) {
-            Log.e("Connectivity Exception", e.getMessage());
-        }
-        return connected;
-    }
-
-    public boolean isOnline() {
-
-        Runtime runtime = Runtime.getRuntime();
-        try {
-            Process ipProcess = runtime.exec("/system/bin/ping -c 1 8.8.8.8");
-            int     exitValue = ipProcess.waitFor();
-            return (exitValue == 0);
-        }
-        catch (IOException | InterruptedException e)          { e.printStackTrace(); }
-
-        return false;
-    }
-
-    public class CheckAddress extends AsyncTask<Void, Void, Void> {
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            waitProgress.show(getSupportFragmentManager(),"WaitBar");
-            waitProgress.setCancelable(false);
-        }
-
-        @Override
-        protected Void doInBackground(Void... voids) {
-            if (isConnected() && isOnline()) {
-                conn = true;
-                getAddress(Double.parseDouble(lat),Double.parseDouble(lon));
-            }
-            else {
-                conn = false;
-                message = "Not Connected";
-            }
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Void aVoid) {
-            super.onPostExecute(aVoid);
-
-            if (conn) {
-                conn = false;
-
-//                new Check().execute();
-                giveAttendance();
-            }
-            else {
-                waitProgress.dismiss();
-                AlertDialog dialog = new AlertDialog.Builder(AttendanceGive.this)
-                        .setMessage("Please Check Your Internet Connection")
-                        .setPositiveButton("Retry", null)
-                        .setNegativeButton("Cancel",null)
-                        .show();
-
-                dialog.setCancelable(false);
-                dialog.setCanceledOnTouchOutside(false);
-                Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-                positive.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-
-                        new CheckAddress().execute();
+                        checkAddress();
                         dialog.dismiss();
-                    }
-                });
-                Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-                negative.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        dialog.dismiss();
-
-                    }
-                });
-            }
-
-        }
+                    });
+                    Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
+                    negative.setOnClickListener(v -> dialog.dismiss());
+                }
+            });
+        }).start();
     }
+
+//    public boolean isConnected() {
+//        boolean connected = false;
+//        boolean isMobile = false;
+//        try {
+//            ConnectivityManager cm = (ConnectivityManager) getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
+//            @SuppressLint("MissingPermission") NetworkInfo nInfo = cm.getActiveNetworkInfo();
+//            connected = nInfo != null && nInfo.isAvailable() && nInfo.isConnected();
+//            return connected;
+//        } catch (Exception e) {
+//            Log.e("Connectivity Exception", e.getMessage());
+//        }
+//        return connected;
+//    }
+//
+//    public boolean isOnline() {
+//
+//        Runtime runtime = Runtime.getRuntime();
+//        try {
+//            Process ipProcess = runtime.exec("/system/bin/ping -c 1 8.8.8.8");
+//            int     exitValue = ipProcess.waitFor();
+//            return (exitValue == 0);
+//        }
+//        catch (IOException | InterruptedException e)          { logger.log(Level.WARNING,e.getMessage(),e); }
+//
+//        return false;
+//    }
+//
+//    public class CheckAddress extends AsyncTask<Void, Void, Void> {
+//
+//        @Override
+//        protected void onPreExecute() {
+//            super.onPreExecute();
+//            waitProgress.show(getSupportFragmentManager(),"WaitBar");
+//            waitProgress.setCancelable(false);
+//        }
+//
+//        @Override
+//        protected Void doInBackground(Void... voids) {
+//            if (isConnected() && isOnline()) {
+//                conn = true;
+//                getAddress(Double.parseDouble(lat),Double.parseDouble(lon));
+//            }
+//            else {
+//                conn = false;
+//                message = "Not Connected";
+//            }
+//            return null;
+//        }
+//
+//        @Override
+//        protected void onPostExecute(Void aVoid) {
+//            super.onPostExecute(aVoid);
+//
+//            if (conn) {
+//                conn = false;
+//
+////                new Check().execute();
+//                giveAttendance();
+//            }
+//            else {
+//                waitProgress.dismiss();
+//                AlertDialog dialog = new AlertDialog.Builder(AttendanceGive.this)
+//                        .setMessage("Please Check Your Internet Connection")
+//                        .setPositiveButton("Retry", null)
+//                        .setNegativeButton("Cancel",null)
+//                        .show();
+//
+//                dialog.setCancelable(false);
+//                dialog.setCanceledOnTouchOutside(false);
+//                Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
+//                positive.setOnClickListener(new View.OnClickListener() {
+//                    @Override
+//                    public void onClick(View v) {
+//
+//                        new CheckAddress().execute();
+//                        dialog.dismiss();
+//                    }
+//                });
+//                Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
+//                negative.setOnClickListener(new View.OnClickListener() {
+//                    @Override
+//                    public void onClick(View v) {
+//                        dialog.dismiss();
+//
+//                    }
+//                });
+//            }
+//
+//        }
+//    }
 
 //    public class Check extends AsyncTask<Void, Void, Void> {
 //
@@ -1047,7 +1033,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
 //
 //            //   Toast.makeText(MainActivity.this, ""+e,Toast.LENGTH_LONG).show();
 //            Log.i("ERRRRR", e.getLocalizedMessage());
-//            e.printStackTrace();
+//            logger.log(Level.WARNING,e.getMessage(),e);
 //        }
 //    }
 
@@ -1055,7 +1041,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
         conn = false;
         connected = false;
 
-        String attendaceUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/giveAttendance";
+        String attendaceUrl = api_url_front + "attendance/giveAttendance";
 
         RequestQueue requestQueue = Volley.newRequestQueue(AttendanceGive.this);
 
@@ -1074,18 +1060,18 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                 updateLayout();
             }
             catch (JSONException e) {
-                e.printStackTrace();
+                logger.log(Level.WARNING,e.getMessage(),e);
                 connected = false;
                 updateLayout();
             }
         }, error -> {
-            error.printStackTrace();
+            logger.log(Level.WARNING,error.getMessage(),error);
             conn = false;
             connected = false;
             updateLayout();
         }) {
             @Override
-            public Map<String, String> getHeaders() throws AuthFailureError {
+            public Map<String, String> getHeaders() {
                 Map<String, String> headers = new HashMap<>();
                 headers.put("P_EMP_ID",emp_id);
                 headers.put("P_PUNCH_TIME",ts.toString());
@@ -1115,7 +1101,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                 editor.putString(timeKey,ss);
                 editor.apply();
                 editor.commit();
-                String puncher = "";
+                String puncher;
                 if (address.isEmpty()) {
 
                     address = "No Address found for ("+lat+", "+lon+")";
@@ -1128,16 +1114,18 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                 currLoc.setVisibility(View.VISIBLE);
 
                 if (tracking_flag == 1) {
-                    if (isMyServiceRunning(Service.class)) {
+                    if (isMyServiceRunning()) {
                         System.out.println("Service Stopped");
 
                         stopService();
-                        nameOfCheckIN.setText("PUNCH & START TRACKER");
+                        String tt = "PUNCH & START TRACKER";
+                        nameOfCheckIN.setText(tt);
                     } else {
                         System.out.println("Service Started");
 
                         startService();
-                        nameOfCheckIN.setText("PUNCH & STOP TRACKER");
+                        String tt = "PUNCH & STOP TRACKER";
+                        nameOfCheckIN.setText(tt);
                     }
                 }
 
@@ -1145,13 +1133,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                 AlertDialog.Builder builder = new AlertDialog.Builder(AttendanceGive.this);
                 builder
                         .setMessage("Your Attendance is Recorded!")
-                        .setPositiveButton("OK", new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which) {
-                                dialog.dismiss();
-
-                            }
-                        });
+                        .setPositiveButton("OK", (dialog, which) -> dialog.dismiss());
 
                 AlertDialog alert = builder.create();
                 alert.show();
@@ -1169,22 +1151,13 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                 dialog.setCancelable(false);
                 dialog.setCanceledOnTouchOutside(false);
                 Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-                positive.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
+                positive.setOnClickListener(v -> {
 
-                        new CheckAddress().execute();
-                        dialog.dismiss();
-                    }
+                    checkAddress();
+                    dialog.dismiss();
                 });
                 Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-                negative.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        dialog.dismiss();
-
-                    }
-                });
+                negative.setOnClickListener(v -> dialog.dismiss());
             }
         }
         else {
@@ -1197,21 +1170,13 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
             dialog.setCancelable(false);
             dialog.setCanceledOnTouchOutside(false);
             Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-            positive.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
+            positive.setOnClickListener(v -> {
 
-                    new CheckAddress().execute();
-                    dialog.dismiss();
-                }
+                checkAddress();
+                dialog.dismiss();
             });
             Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-            negative.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    dialog.dismiss();
-                }
-            });
+            negative.setOnClickListener(v -> dialog.dismiss());
         }
     }
 
@@ -1244,7 +1209,7 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
 //
 //            //   Toast.makeText(MainActivity.this, ""+e,Toast.LENGTH_LONG).show();
 //            Log.i("ERRRRR", e.getLocalizedMessage());
-//            e.printStackTrace();
+//            logger.log(Level.WARNING,e.getMessage(),e);
 //        }
 //    }
 
@@ -1254,7 +1219,9 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
         conn = false;
         connected = false;
 
-        String offLocationUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/getOffLatLong/"+emp_id+"";
+        areaLists = new ArrayList<>();
+
+        String offLocationUrl = api_url_front + "attendance/getNewOffLatLong?emp_id="+emp_id;
 
         RequestQueue requestQueue = Volley.newRequestQueue(AttendanceGive.this);
 
@@ -1268,21 +1235,33 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                     JSONArray array = new JSONArray(items);
                     for (int i = 0; i < array.length(); i++) {
                         JSONObject offLocInfo = array.getJSONObject(i);
-                        officeLatitude = offLocInfo.getString("coa_latitude").equals("null") ? null : offLocInfo.getString("coa_latitude");
-                        officeLongitude = offLocInfo.getString("coa_longitude").equals("null") ? null : offLocInfo.getString("coa_longitude");
-                        coverage = offLocInfo.getString("coa_coverage").equals("null") ? null : offLocInfo.getString("coa_coverage");
+                        String coa_latitude = offLocInfo.getString("coa_latitude")
+                                .equals("null") ? null : offLocInfo.getString("coa_latitude");
+                        String coa_longitude = offLocInfo.getString("coa_longitude")
+                                .equals("null") ? null : offLocInfo.getString("coa_longitude");
+                        String coa_coverage = offLocInfo.getString("coa_coverage")
+                                .equals("null") ? null : offLocInfo.getString("coa_coverage");
+                        String co_id = offLocInfo.getString("coa_id")
+                                .equals("null") ? null : offLocInfo.getString("coa_id");
+                        String code = offLocInfo.getString("machine_code")
+                                .equals("null") ? null : offLocInfo.getString("machine_code");
+                        String can_give = offLocInfo.getString("can_give")
+                                .equals("null") ? "0" : offLocInfo.getString("can_give");
+
+                        areaLists.add(new AreaList(coa_latitude,coa_longitude,coa_coverage,co_id,code,can_give.equals("1")));
+
                     }
                 }
                 connected = true;
                 updateInfo();
             }
             catch (JSONException e) {
-                e.printStackTrace();
+                logger.log(Level.WARNING,e.getMessage(),e);
                 connected = false;
                 updateInfo();
             }
         }, error -> {
-            error.printStackTrace();
+            logger.log(Level.WARNING,error.getMessage(),error);
             conn = false;
             connected = false;
             updateInfo();
@@ -1310,21 +1289,15 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
                 dialog.setCancelable(false);
                 dialog.setCanceledOnTouchOutside(false);
                 Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-                positive.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
+                positive.setOnClickListener(v -> {
 
-                        getOfficeLocation();
-                        dialog.dismiss();
-                    }
+                    getOfficeLocation();
+                    dialog.dismiss();
                 });
                 Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-                negative.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        dialog.dismiss();
-                        finish();
-                    }
+                negative.setOnClickListener(v -> {
+                    dialog.dismiss();
+                    finish();
                 });
             }
         }
@@ -1338,21 +1311,15 @@ public class AttendanceGive extends AppCompatActivity implements OnMapReadyCallb
             dialog.setCancelable(false);
             dialog.setCanceledOnTouchOutside(false);
             Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-            positive.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
+            positive.setOnClickListener(v -> {
 
-                    getOfficeLocation();
-                    dialog.dismiss();
-                }
+                getOfficeLocation();
+                dialog.dismiss();
             });
             Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-            negative.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    dialog.dismiss();
-                    finish();
-                }
+            negative.setOnClickListener(v -> {
+                dialog.dismiss();
+                finish();
             });
         }
     }
