@@ -2,12 +2,7 @@ package ttit.com.shuvo.ikglhrm.directoryBook;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.res.ColorStateList;
-import android.graphics.Color;
 import android.net.Uri;
-import android.text.SpannableString;
-import android.text.Spanned;
-import android.text.style.ForegroundColorSpan;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -19,7 +14,6 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
@@ -41,8 +35,7 @@ public class PhoneAdapter extends RecyclerView.Adapter<PhoneAdapter.PhoneHolder>
     @Override
     public PhoneHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View v = LayoutInflater.from(myContext).inflate(R.layout.phone_item, parent, false);
-        PhoneHolder ammvh = new PhoneHolder(v);
-        return ammvh;
+        return new PhoneHolder(v);
     }
 
     @Override
@@ -54,7 +47,8 @@ public class PhoneAdapter extends RecyclerView.Adapter<PhoneAdapter.PhoneHolder>
         String phone = foodOptionCategoryList.getPhone();
 
         if (phone == null) {
-            holder.phoneNumber.setText("No Number Found");
+            String nnft = "No Number Found";
+            holder.phoneNumber.setText(nnft);
             holder.phoneClick.setVisibility(View.GONE);
         } else {
             char[] chars = phone.toCharArray();
@@ -67,7 +61,8 @@ public class PhoneAdapter extends RecyclerView.Adapter<PhoneAdapter.PhoneHolder>
             }
 
             if (!found) {
-                holder.phoneNumber.setText("No Number Found");
+                String nnft = "No Number Found";
+                holder.phoneNumber.setText(nnft);
                 holder.phoneClick.setVisibility(View.GONE);
             } else {
                 holder.phoneClick.setVisibility(View.VISIBLE);
@@ -83,7 +78,7 @@ public class PhoneAdapter extends RecyclerView.Adapter<PhoneAdapter.PhoneHolder>
         return phoneLists.size();
     }
 
-    public class PhoneHolder extends RecyclerView.ViewHolder {
+    public static class PhoneHolder extends RecyclerView.ViewHolder {
 
         TextView phoneNumber;
         ImageView phoneClick;
@@ -96,16 +91,14 @@ public class PhoneAdapter extends RecyclerView.Adapter<PhoneAdapter.PhoneHolder>
             phoneClick = itemView.findViewById(R.id.phone_click);
 
 
-            phoneClick.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    Log.i("Name", phoneNumber.getText().toString());
-                    String number = phoneNumber.getText().toString();
-                    //  Log.i("ID", foodOptionCategoryLists.get(getAdapterPosition()).getFoodCategoryID());
+            phoneClick.setOnClickListener(v -> {
+                Log.i("Name", phoneNumber.getText().toString());
+                String number = phoneNumber.getText().toString();
+                //  Log.i("ID", foodOptionCategoryLists.get(getAdapterPosition()).getFoodCategoryID());
 //                    fID = foodOptionCategoryLists.get(getAdapterPosition()).getFoodCategoryID();
 //                    fname = foodCategoryName.getText().toString();
 //                    fOpID = foodOptionCategoryLists.get(getAdapterPosition()).getOptionID();
-                    AppCompatActivity activity = (AppCompatActivity) v.getContext();
+                AppCompatActivity activity = (AppCompatActivity) v.getContext();
 //                    if (fname.equals("FLAVOR")) {
 //                        OptionsFlavorDialogue optionsFlavorDialogue = new OptionsFlavorDialogue();
 //                        optionsFlavorDialogue.show(activity.getSupportFragmentManager(),"FlavorOption");
@@ -117,36 +110,25 @@ public class PhoneAdapter extends RecyclerView.Adapter<PhoneAdapter.PhoneHolder>
 ////                    OptionsMenu optionsMenu = (OptionsMenu) context;
 ////                    FragmentManager fm = ((FragmentActivity)myContext).getSupportFragmentManager();
 //                        optionsDialogoue.show(activity.getSupportFragmentManager(),"sww");
-                    AlertDialog dialog = new AlertDialog.Builder(activity)
-                            .setMessage("Do you want to make a call to "+number+" ?")
-                            .setPositiveButton("Yes", null)
-                            .setNegativeButton("No",null)
-                            .show();
+                AlertDialog dialog = new AlertDialog.Builder(activity)
+                        .setMessage("Do you want to make a call to "+number+" ?")
+                        .setPositiveButton("Yes", null)
+                        .setNegativeButton("No",null)
+                        .show();
 
-                    dialog.setCancelable(false);
-                    dialog.setCanceledOnTouchOutside(false);
-                    Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
-                    positive.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
+                dialog.setCancelable(false);
+                dialog.setCanceledOnTouchOutside(false);
+                Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
+                positive.setOnClickListener(v1 -> {
 
-                            dialog.dismiss();
-                            Intent callIntent = new Intent(Intent.ACTION_DIAL);
-                            callIntent.setData(Uri.parse("tel:"+number));
-                            activity.startActivity(callIntent);
+                    dialog.dismiss();
+                    Intent callIntent = new Intent(Intent.ACTION_DIAL);
+                    callIntent.setData(Uri.parse("tel:"+number));
+                    activity.startActivity(callIntent);
 
-                        }
-                    });
-                    Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
-                    negative.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            dialog.dismiss();
-                        }
-                    });
-
-
-                }
+                });
+                Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
+                negative.setOnClickListener(v12 -> dialog.dismiss());
 
 
             });
