@@ -5,7 +5,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.cardview.widget.CardView;
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
@@ -14,7 +13,6 @@ import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.github.dewinjm.monthyearpicker.MonthFormat;
-import com.github.dewinjm.monthyearpicker.MonthYearPickerDialog;
 import com.github.dewinjm.monthyearpicker.MonthYearPickerDialogFragment;
 import com.github.mikephil.charting.charts.BarChart;
 import com.github.mikephil.charting.components.AxisBase;
@@ -32,6 +30,8 @@ import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import ttit.com.shuvo.ikglhrm.R;
 import ttit.com.shuvo.ikglhrm.WaitProgress;
@@ -39,9 +39,8 @@ import ttit.com.shuvo.ikglhrm.leaveAll.leaveApplication.AllLeaveApplication;
 import ttit.com.shuvo.ikglhrm.leaveAll.leaveBalance.LeaveBalance;
 import ttit.com.shuvo.ikglhrm.leaveAll.leaveStatus.LeaveStatus;
 
-import static ttit.com.shuvo.ikglhrm.Login.CompanyName;
-import static ttit.com.shuvo.ikglhrm.Login.SoftwareName;
 import static ttit.com.shuvo.ikglhrm.Login.userInfoLists;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -49,30 +48,18 @@ import org.json.JSONObject;
 
 public class Leave extends AppCompatActivity {
 
-    TextView softName;
-    TextView compName;
-
     CardView leaveBalance;
     CardView leaveApplication;
     CardView leaveStatus;
 
-    Button back;
-
-//    PieChart pieChart;
     TextView refresh;
 
     WaitProgress waitProgress = new WaitProgress();
-//    private String message = null;
     private Boolean conn = false;
     private Boolean connected = false;
 
-//    private Connection connection;
-
     String emp_id = "";
     String leaveDate = "";
-
-//    ArrayList<PieEntry> NoOfEmp;
-//    ArrayList<SalaryMonthList> dataValue;
 
     ArrayList<BarEntry> balance;
     ArrayList<BarEntry> earn;
@@ -80,72 +67,44 @@ public class Leave extends AppCompatActivity {
 
     BarChart chart;
 
+    String total_leave = "";
+    String used_leave = "";
+    String remain_leave = "";
+
+    TextView totalLeave;
+    TextView usedLeave;
+    TextView remainLeave;
+
+    Logger logger = Logger.getLogger(Leave.class.getName());
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-//        getWindow().requestFeature(Window.FEATURE_NO_TITLE);
-//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-//            Window w = getWindow();
-//            //w.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
-//        }
-//        if (Build.VERSION.SDK_INT < 16) {
-//            getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-//                    WindowManager.LayoutParams.FLAG_FULLSCREEN);
-//        }
-//        View decorView = getWindow().getDecorView();
-//// Hide the status bar.
-//        int uiOptions = View.SYSTEM_UI_FLAG_FULLSCREEN;
-//        decorView.setSystemUiVisibility(uiOptions);
-
         setContentView(R.layout.activity_leave);
-
-        softName = findViewById(R.id.name_of_soft_leave);
-        compName = findViewById(R.id.name_of_company_leave);
 
         leaveBalance = findViewById(R.id.leave_balance_sheet);
         leaveApplication = findViewById(R.id.leave_application);
         leaveStatus = findViewById(R.id.leave_status);
 
-        back = findViewById(R.id.leave_back);
+        totalLeave = findViewById(R.id.total_leave_count_leave);
+        usedLeave = findViewById(R.id.used_leave_leave);
+        remainLeave = findViewById(R.id.remaining_leave_leave);
 
-//        pieChart = findViewById(R.id.piechart);
         refresh = findViewById(R.id.refresh_graph_leave);
 
-        softName.setText(CompanyName);
-        compName.setText(SoftwareName);
-
-//        dataValue = new ArrayList<>();
-
-        back.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
+        leaveBalance.setOnClickListener(v -> {
+            Intent intent = new Intent(Leave.this, LeaveBalance.class);
+            startActivity(intent);
         });
 
-        leaveBalance.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(Leave.this, LeaveBalance.class);
-                startActivity(intent);
-            }
+        leaveApplication.setOnClickListener(v -> {
+            Intent intent = new Intent(Leave.this, AllLeaveApplication.class);
+            startActivity(intent);
         });
 
-        leaveApplication.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(Leave.this, AllLeaveApplication.class);
-                startActivity(intent);
-            }
-        });
-
-        leaveStatus.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(Leave.this, LeaveStatus.class);
-                startActivity(intent);
-            }
+        leaveStatus.setOnClickListener(v -> {
+            Intent intent = new Intent(Leave.this, LeaveStatus.class);
+            startActivity(intent);
         });
 
         emp_id = userInfoLists.get(0).getEmp_id();
@@ -155,9 +114,10 @@ public class Leave extends AppCompatActivity {
         SimpleDateFormat df = new SimpleDateFormat("dd-MMM-yy", Locale.ENGLISH);
         SimpleDateFormat mmm = new SimpleDateFormat("MMMM", Locale.ENGLISH);
 
-        String monnnn = mmm.format(c);
+        String monnnn = mmm.format(c).toUpperCase();
 
-        refresh.setText("Month: "+ monnnn);
+        String rt = "Month: "+ monnnn;
+        refresh.setText(rt);
 
         leaveDate = df.format(c);
 
@@ -185,243 +145,173 @@ public class Leave extends AppCompatActivity {
 
         chart.getAxisRight().setEnabled(false);
 
-//
-//        NoOfEmp = new ArrayList<>();
-//
-//        NoOfEmp.add(new PieEntry(945f, "2008", 0));
-//        NoOfEmp.add(new PieEntry(1040f,"2009", 1));
-//        NoOfEmp.add(new PieEntry(1133f,"2010", 2));
-//        NoOfEmp.add(new PieEntry(1240f,"2011", 3));
-//        NoOfEmp.add(new PieEntry(1369f,"2012", 4));
-//        NoOfEmp.add(new PieEntry(1487f,"2013", 5));
-//        NoOfEmp.add(new PieEntry(1501f,"2014", 6));
-//        NoOfEmp.add(new PieEntry(1645f,"2015", 7));
-//        NoOfEmp.add(new PieEntry(1578f,"2016", 8));
-//        NoOfEmp.add(new PieEntry(1695f,"2017", 9));
-//
-//        PieDataSet dataSet = new PieDataSet(NoOfEmp, "");
+        getLeaveData();
 
+        refresh.setOnClickListener(v -> {
 
+            Date c1 = Calendar.getInstance().getTime();
 
+            String formattedYear;
+            String monthValue;
+            String lastformattedYear;
+            String lastdateView;
 
+            SimpleDateFormat df1 = new SimpleDateFormat("yyyy", Locale.ENGLISH);
+            SimpleDateFormat sdf = new SimpleDateFormat("MM",Locale.ENGLISH);
 
-//        pieChart.setCenterText("Leave Balance");
-//        pieChart.setDrawEntryLabels(true);
-//        pieChart.setCenterTextSize(14);
-//        pieChart.setHoleRadius(40);
-//        pieChart.setTransparentCircleRadius(40);
+            formattedYear = df1.format(c1);
+            monthValue = sdf.format(c1);
+            int nowMonNumb = Integer.parseInt(monthValue);
+            nowMonNumb = nowMonNumb - 1;
+            int lastMonNumb = nowMonNumb - 5;
 
-        // eta bad
-//        pieChart.setTransparentCircleColor(Color.RED);
-//        pieChart.setTransparentCircleAlpha(50);
-//        pieChart.setCenterTextRadiusPercent(30);
+            if (lastMonNumb < 0) {
+                lastMonNumb = lastMonNumb + 12;
+                int formatY = Integer.parseInt(formattedYear);
+                formatY = formatY - 1;
+                lastformattedYear = String.valueOf(formatY);
+            } else {
+                lastformattedYear = formattedYear;
+            }
 
+            Date today = new Date();
 
-//        pieChart.setEntryLabelTextSize(11);
-//        pieChart.setEntryLabelColor(Color.DKGRAY);
-//        pieChart.getDescription().setEnabled(false);
+            Calendar calendar1 = Calendar.getInstance();
+            calendar1.setTime(today);
 
+            calendar1.add(Calendar.MONTH, 1);
+            calendar1.set(Calendar.DAY_OF_MONTH, 1);
+            calendar1.add(Calendar.DATE, -1);
 
-//        Legend l = pieChart.getLegend();
-//        l.setVerticalAlignment(Legend.LegendVerticalAlignment.CENTER);
-//        l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
-//        l.setOrientation(Legend.LegendOrientation.VERTICAL);
-//        l.setXOffset(20);
-//        l.setTextSize(12);
+            Date lastDayOfMonth = calendar1.getTime();
 
-        // eta bad
-        //l.setYOffset(50);
+            SimpleDateFormat sdff = new SimpleDateFormat("dd",Locale.ENGLISH);
+            lastdateView = sdff.format(lastDayOfMonth);
 
-//        l.setWordWrapEnabled(false);
-//        l.setDrawInside(false);
-//        l.setYOffset(5f);
-//        final int[] piecolors = new int[]{
-//                Color.rgb(129, 236, 236),
-//                Color.rgb(85, 239, 196),
-//                Color.rgb(116, 185, 255),
-//                Color.rgb(162, 155, 254),
-//                Color.rgb(223, 230, 233),
-//                Color.rgb(255, 234, 167),
-//                Color.rgb(250, 177, 160),
-//                Color.rgb(255, 118, 117),
-//                Color.rgb(253, 121, 168),
-//                Color.rgb(96, 163, 188)};
-//        pieChart.animateXY(1500, 1500);
-
-//        PieData data = new PieData(dataSet);
-//
-//        dataSet.setHighlightEnabled(true);
-//        dataSet.setValueTextSize(12);
-//        dataSet.setValueTextColor(Color.BLACK);
-//        dataSet.setColors(ColorTemplate.createColors(piecolors));
-
-//        pieChart.setUsePercentValues(true);
-
-
-//        pieChart.setData(data);
-//        pieChart.invalidate();
-
-//        new Check().execute();
-        getLeaveGraph();
-
-
-        refresh.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                Date c = Calendar.getInstance().getTime();
-
-                String formattedYear = "";
-                String monthValue = "";
-                String lastformattedYear = "";
-                String lastdateView = "";
-
-                SimpleDateFormat df = new SimpleDateFormat("yyyy", Locale.ENGLISH);
-                SimpleDateFormat sdf = new SimpleDateFormat("MM",Locale.ENGLISH);
-
-                formattedYear = df.format(c);
-                monthValue = sdf.format(c);
-                int nowMonNumb = Integer.parseInt(monthValue);
-                nowMonNumb = nowMonNumb - 1;
-                int lastMonNumb = nowMonNumb - 5;
-
-                if (lastMonNumb < 0) {
-                    lastMonNumb = lastMonNumb + 12;
-                    int formatY = Integer.parseInt(formattedYear);
-                    formatY = formatY - 1;
-                    lastformattedYear = String.valueOf(formatY);
-                } else {
-                    lastformattedYear = formattedYear;
+            int yearSelected;
+            int monthSelected;
+            MonthFormat monthFormat = MonthFormat.LONG;
+            String customTitle = "Select Month";
+// Use the calendar for create ranges
+            Calendar calendar = Calendar.getInstance();
+            if (!leaveDate.isEmpty()) {
+                SimpleDateFormat myf = new SimpleDateFormat("dd-MMM-yy", Locale.ENGLISH);
+                Date md = null;
+                try {
+                    md = myf.parse(leaveDate);
+                } catch (ParseException e) {
+                    logger.log(Level.WARNING, e.getMessage(), e);
                 }
 
-                Date today = new Date();
+                if (md != null) {
+                    calendar.setTime(md);
+                }
+            }
+            yearSelected = calendar.get(Calendar.YEAR);
+            monthSelected = calendar.get(Calendar.MONTH);
+            calendar.clear();
+            calendar.set(Integer.parseInt(lastformattedYear), lastMonNumb, 1); // Set minimum date to show in dialog
+            long minDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
 
-                Calendar calendar1 = Calendar.getInstance();
-                calendar1.setTime(today);
-
-                calendar1.add(Calendar.MONTH, 1);
-                calendar1.set(Calendar.DAY_OF_MONTH, 1);
-                calendar1.add(Calendar.DATE, -1);
-
-                Date lastDayOfMonth = calendar1.getTime();
-
-                SimpleDateFormat sdff = new SimpleDateFormat("dd",Locale.ENGLISH);
-                lastdateView = sdff.format(lastDayOfMonth);
-
-                int yearSelected;
-                int monthSelected;
-                MonthFormat monthFormat = MonthFormat.LONG;
-                String customTitle = "Select Month";
-// Use the calendar for create ranges
-                Calendar calendar = Calendar.getInstance();
-                yearSelected = calendar.get(Calendar.YEAR);
-                monthSelected = calendar.get(Calendar.MONTH);
-                calendar.clear();
-                calendar.set(Integer.parseInt(lastformattedYear), lastMonNumb, 1); // Set minimum date to show in dialog
-                long minDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
-
-                calendar.clear();
-                calendar.set(Integer.parseInt(formattedYear), nowMonNumb, Integer.parseInt(lastdateView)); // Set maximum date to show in dialog
-                long maxDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
+            calendar.clear();
+            calendar.set(Integer.parseInt(formattedYear), nowMonNumb, Integer.parseInt(lastdateView)); // Set maximum date to show in dialog
+            long maxDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
 
 // Create instance with date ranges values
-                MonthYearPickerDialogFragment dialogFragment =  MonthYearPickerDialogFragment
-                        .getInstance(monthSelected, yearSelected, minDate, maxDate, customTitle, monthFormat);
+            MonthYearPickerDialogFragment dialogFragment =  MonthYearPickerDialogFragment
+                    .getInstance(monthSelected, yearSelected, minDate, maxDate, customTitle, monthFormat);
 
 
 
-                dialogFragment.show(getSupportFragmentManager(), null);
+            dialogFragment.show(getSupportFragmentManager(), null);
 
-                dialogFragment.setOnDateSetListener(new MonthYearPickerDialog.OnDateSetListener() {
-                    @Override
-                    public void onDateSet(int year, int monthOfYear) {
-                        System.out.println(year);
-                        System.out.println(monthOfYear);
+            dialogFragment.setOnDateSetListener((year, monthOfYear) -> {
+                System.out.println(year);
+                System.out.println(monthOfYear);
 
-                        int month = monthOfYear + 1;
-                        String monthName = "";
-                        String mon = "";
-                        String yearName = "";
+                int month = monthOfYear + 1;
+                String monthName = "";
+                String mon = "";
+                String yearName;
 
-                        if (month == 1) {
-                            monthName = "JANUARY";
-                            mon = "JAN";
-                        } else if (month == 2) {
-                            monthName = "FEBRUARY";
-                            mon = "FEB";
-                        } else if (month == 3) {
-                            monthName = "MARCH";
-                            mon = "MAR";
-                        } else if (month == 4) {
-                            monthName = "APRIL";
-                            mon = "APR";
-                        } else if (month == 5) {
-                            monthName = "MAY";
-                            mon = "MAY";
-                        } else if (month == 6) {
-                            monthName = "JUNE";
-                            mon = "JUN";
-                        } else if (month == 7) {
-                            monthName = "JULY";
-                            mon = "JUL";
-                        } else if (month == 8) {
-                            monthName = "AUGUST";
-                            mon = "AUG";
-                        } else if (month == 9) {
-                            monthName = "SEPTEMBER";
-                            mon = "SEP";
-                        } else if (month == 10) {
-                            monthName = "OCTOBER";
-                            mon = "OCT";
-                        } else if (month == 11) {
-                            monthName = "NOVEMBER";
-                            mon = "NOV";
-                        } else if (month == 12) {
-                            monthName = "DECEMBER";
-                            mon = "DEC";
-                        }
+                if (month == 1) {
+                    monthName = "JANUARY";
+                    mon = "JAN";
+                } else if (month == 2) {
+                    monthName = "FEBRUARY";
+                    mon = "FEB";
+                } else if (month == 3) {
+                    monthName = "MARCH";
+                    mon = "MAR";
+                } else if (month == 4) {
+                    monthName = "APRIL";
+                    mon = "APR";
+                } else if (month == 5) {
+                    monthName = "MAY";
+                    mon = "MAY";
+                } else if (month == 6) {
+                    monthName = "JUNE";
+                    mon = "JUN";
+                } else if (month == 7) {
+                    monthName = "JULY";
+                    mon = "JUL";
+                } else if (month == 8) {
+                    monthName = "AUGUST";
+                    mon = "AUG";
+                } else if (month == 9) {
+                    monthName = "SEPTEMBER";
+                    mon = "SEP";
+                } else if (month == 10) {
+                    monthName = "OCTOBER";
+                    mon = "OCT";
+                } else if (month == 11) {
+                    monthName = "NOVEMBER";
+                    mon = "NOV";
+                } else if (month == 12) {
+                    monthName = "DECEMBER";
+                    mon = "DEC";
+                }
 
-                        yearName  = String.valueOf(year);
-                        yearName = yearName.substring(yearName.length()-2);
+                yearName  = String.valueOf(year);
+                yearName = yearName.substring(yearName.length()-2);
 
 
 
-                        leaveDate = "01-"+mon+"-"+yearName;
-                        //selected_date = "01-"+mon+"-"+yearName;
-                        refresh.setText("Month: "+monthName);
+                leaveDate = "01-"+mon+"-"+yearName;
+                //selected_date = "01-"+mon+"-"+yearName;
+                String rt1 = "Month: "+monthName;
+                refresh.setText(rt1);
 
-                        SimpleDateFormat sss = new SimpleDateFormat("dd-MMM-yyyy", Locale.ENGLISH);
+                SimpleDateFormat sss = new SimpleDateFormat("dd-MMM-yyyy", Locale.ENGLISH);
 
-                        Date today = null;
-                        try {
-                            today = sss.parse(leaveDate);
-                        } catch (ParseException e) {
-                            e.printStackTrace();
-                        }
+                Date today1 = null;
+                try {
+                    today1 = sss.parse(leaveDate);
+                } catch (ParseException e) {
+                    logger.log(Level.WARNING, e.getMessage(), e);
+                }
 
-                        Calendar calendar1 = Calendar.getInstance();
-                        if (today != null) {
-                            calendar1.setTime(today);
-                            calendar1.add(Calendar.MONTH, 1);
-                            calendar1.set(Calendar.DAY_OF_MONTH, 1);
-                            calendar1.add(Calendar.DATE, -1);
+                Calendar calendar11 = Calendar.getInstance();
+                if (today1 != null) {
+                    calendar11.setTime(today1);
+                    calendar11.add(Calendar.MONTH, 1);
+                    calendar11.set(Calendar.DAY_OF_MONTH, 1);
+                    calendar11.add(Calendar.DATE, -1);
 
-                            Date lastDayOfMonth = calendar1.getTime();
+                    Date lastDayOfMonth1 = calendar11.getTime();
 
-                            SimpleDateFormat sdff = new SimpleDateFormat("dd",Locale.ENGLISH);
-                            String llll = sdff.format(lastDayOfMonth);
-                            leaveDate =  llll+ "-" + mon +"-"+ yearName;
-                        }
-                        System.out.println(leaveDate);
+                    SimpleDateFormat sdff1 = new SimpleDateFormat("dd",Locale.ENGLISH);
+                    String llll = sdff1.format(lastDayOfMonth1);
+                    leaveDate =  llll+ "-" + mon +"-"+ yearName;
+                }
+                System.out.println(leaveDate);
 
 
 
 //                        new Check().execute();
-                        getLeaveGraph();
-                    }
-                });
+                getLeaveGraph();
+            });
 
-            }
         });
     }
 
@@ -447,7 +337,7 @@ public class Leave extends AppCompatActivity {
 //            int     exitValue = ipProcess.waitFor();
 //            return (exitValue == 0);
 //        }
-//        catch (IOException | InterruptedException e)          { e.printStackTrace(); }
+//        catch (IOException | InterruptedException e)          { logger.log(Level.WARNING, e.getMessage(), e); }
 //
 //        return false;
 //    }
@@ -700,9 +590,223 @@ public class Leave extends AppCompatActivity {
 //
 //            //   Toast.makeText(MainActivity.this, ""+e,Toast.LENGTH_LONG).show();
 //            Log.i("ERRRRR", e.getLocalizedMessage());
-//            e.printStackTrace();
+//            logger.log(Level.WARNING, e.getMessage(), e);
 //        }
 //    }
+
+    public void getLeaveData() {
+        waitProgress.show(getSupportFragmentManager(),"WaitBar");
+        waitProgress.setCancelable(false);
+        connected = false;
+        conn = false;
+
+        balance = new ArrayList<>();
+        earn = new ArrayList<>();
+        shortCode = new ArrayList<>();
+
+        BarData leavedata = new BarData();
+        chart.setData(leavedata);
+        chart.getData().clearValues();
+        chart.notifyDataSetChanged();
+        chart.clear();
+        chart.invalidate();
+
+        total_leave = "";
+        used_leave = "";
+        remain_leave = "";
+
+        String leaveDataUrl = api_url_front + "dashboard/getLeaveData/"+emp_id+"/"+leaveDate;
+        String leaveCountUrl = api_url_front + "leave/getLeaveCountSum?emp_id="+emp_id+"&now_date="+leaveDate;
+
+        RequestQueue requestQueue = Volley.newRequestQueue(Leave.this);
+
+        StringRequest leaveCountReq = new StringRequest(Request.Method.GET, leaveCountUrl, response -> {
+            conn = true;
+            try {
+                JSONObject jsonObject = new JSONObject(response);
+                String items = jsonObject.getString("items");
+                String count = jsonObject.getString("count");
+                if (!count.equals("0")) {
+                    JSONArray array = new JSONArray(items);
+                    for (int i = 0; i < array.length(); i++) {
+                        JSONObject leaveInfo = array.getJSONObject(i);
+                        total_leave = leaveInfo.getString("tt").equals("null") ? "0" : leaveInfo.getString("tt");
+                        used_leave = leaveInfo.getString("usd").equals("null") ? "0" : leaveInfo.getString("usd");
+                        remain_leave = leaveInfo.getString("rmn").equals("null") ? "0" : leaveInfo.getString("rmn");
+                    }
+                }
+                connected = true;
+                updateLeaveData();
+            }
+            catch (JSONException e) {
+                connected = false;
+                logger.log(Level.WARNING, e.getMessage(), e);
+                updateLeaveData();
+            }
+        }, error -> {
+            conn = false;
+            connected = false;
+            logger.log(Level.WARNING, error.getMessage(), error);
+            updateLeaveData();
+        });
+
+        StringRequest leaveDataReq = new StringRequest(Request.Method.GET, leaveDataUrl, response -> {
+            conn = true;
+            try {
+                JSONObject jsonObject = new JSONObject(response);
+                String items = jsonObject.getString("items");
+                String count = jsonObject.getString("count");
+                if (!count.equals("0")) {
+                    JSONArray array = new JSONArray(items);
+                    for (int i = 0; i < array.length(); i++) {
+                        JSONObject leaveInfo = array.getJSONObject(i);
+//                        String lbem_emp_id = leaveInfo.getString("lbem_emp_id");
+                        String lc_short_code = leaveInfo.getString("lc_short_code");
+                        String balance_all = leaveInfo.getString("balance");
+                        String quantity = leaveInfo.getString("quantity");
+                        if (!quantity.equals("null") && !quantity.equals("NULL")) {
+                            if (!quantity.equals("0")) {
+                                balance.add(new BarEntry(i, Float.parseFloat(balance_all),i));
+                                earn.add(new BarEntry(i, Float.parseFloat(quantity),i));
+                                shortCode.add(lc_short_code);
+                            }
+                        }
+                    }
+                }
+                requestQueue.add(leaveCountReq);
+            }
+            catch (JSONException e) {
+                connected = false;
+                logger.log(Level.WARNING, e.getMessage(), e);
+                updateLeaveData();
+            }
+        }, error -> {
+            conn = false;
+            connected = false;
+            logger.log(Level.WARNING, error.getMessage(), error);
+            updateLeaveData();
+        });
+
+        requestQueue.add(leaveDataReq);
+    }
+
+    private void updateLeaveData() {
+        waitProgress.dismiss();
+        if (conn) {
+            if (connected) {
+
+                totalLeave.setText(total_leave);
+                usedLeave.setText(used_leave);
+                remainLeave.setText(remain_leave);
+
+                if (!balance.isEmpty() && !earn.isEmpty() && !shortCode.isEmpty()) {
+                    XAxis xAxis = chart.getXAxis();
+                    xAxis.setPosition(XAxis.XAxisPosition.BOTTOM);
+                    xAxis.setDrawGridLines(false);
+                    xAxis.setCenterAxisLabels(true);
+                    xAxis.setAxisMinimum(0);
+                    xAxis.setAxisMaximum(balance.size());
+                    xAxis.setGranularity(1);
+
+
+                    BarDataSet set1 = new BarDataSet(earn, "Earned Leave");
+                    set1.setColor(ColorTemplate.VORDIPLOM_COLORS[3]);
+                    set1.setValueFormatter(new ValueFormatter() {
+                        @Override
+                        public String getFormattedValue(float value) {
+                            return String.valueOf((int) Math.floor(value));
+                        }
+                    });
+                    BarDataSet set2 = new BarDataSet(balance, "Leave Balance");
+                    set2.setValueFormatter(new ValueFormatter() {
+                        @Override
+                        public String getFormattedValue(float value) {
+                            return String.valueOf((int) Math.floor(value));
+                        }
+                    });
+                    set2.setColor(ColorTemplate.VORDIPLOM_COLORS[2]);
+
+                    float groupSpace = 0.04f;
+                    float barSpace = 0.02f; // x2 dataset
+                    float barWidth = 0.46f;
+
+                    BarData data = new BarData(set1, set2);
+                    if (earn.size() > 10) {
+                        data.setValueTextSize(8);
+                    }
+                    else if (earn.size() > 5) {
+                        data.setValueTextSize(11);
+                    }
+                    else {
+                        data.setValueTextSize(12);
+                    }
+                    data.setBarWidth(barWidth); // set the width of each bar
+                    chart.animateY(1000);
+                    chart.setData(data);
+                    chart.groupBars(0, groupSpace, barSpace); // perform the "explicit" grouping
+                    chart.invalidate();
+
+                    xAxis.setValueFormatter(new ValueFormatter() {
+                        @Override
+                        public String getAxisLabel(float value, AxisBase axis) {
+                            if (value < 0 || value >= shortCode.size()) {
+                                return null;
+                            } else {
+//                                System.out.println(value);
+//                                System.out.println(axis);
+//                                System.out.println(shortCode.get((int)value));
+                                return (shortCode.get((int) value));
+                            }
+
+                        }
+                    });
+
+                }
+            }
+            else {
+                AlertDialog dialog = new AlertDialog.Builder(Leave.this)
+                        .setMessage("There is a network issue in the server. Please Try later.")
+                        .setPositiveButton("Retry", null)
+                        .setNegativeButton("Cancel", null)
+                        .show();
+
+                dialog.setCancelable(false);
+                dialog.setCanceledOnTouchOutside(false);
+                Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
+                positive.setOnClickListener(v -> {
+                    getLeaveData();
+                    dialog.dismiss();
+                });
+
+                Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
+                negative.setOnClickListener(v -> {
+                    dialog.dismiss();
+                    finish();
+                });
+            }
+        }
+        else {
+            AlertDialog dialog = new AlertDialog.Builder(Leave.this)
+                    .setMessage("Please Check Your Internet Connection")
+                    .setPositiveButton("Retry", null)
+                    .setNegativeButton("Cancel", null)
+                    .show();
+
+            dialog.setCancelable(false);
+            dialog.setCanceledOnTouchOutside(false);
+            Button positive = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
+            positive.setOnClickListener(v -> {
+                getLeaveData();
+                dialog.dismiss();
+            });
+
+            Button negative = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
+            negative.setOnClickListener(v -> {
+                dialog.dismiss();
+                finish();
+            });
+        }
+    }
 
     public void getLeaveGraph() {
         waitProgress.show(getSupportFragmentManager(),"WaitBar");
@@ -714,7 +818,14 @@ public class Leave extends AppCompatActivity {
         earn = new ArrayList<>();
         shortCode = new ArrayList<>();
 
-        String leaveDataUrl = "http://103.56.208.123:8001/apex/ttrams/dashboard/getLeaveData/"+emp_id+"/"+leaveDate+"";
+        BarData leavedata = new BarData();
+        chart.setData(leavedata);
+        chart.getData().clearValues();
+        chart.notifyDataSetChanged();
+        chart.clear();
+        chart.invalidate();
+
+        String leaveDataUrl = api_url_front + "dashboard/getLeaveData/"+emp_id+"/"+leaveDate;
 
         RequestQueue requestQueue = Volley.newRequestQueue(Leave.this);
 
@@ -746,13 +857,13 @@ public class Leave extends AppCompatActivity {
             }
             catch (JSONException e) {
                 connected = false;
-                e.printStackTrace();
+                logger.log(Level.WARNING, e.getMessage(), e);
                 updateLeaveGraph();
             }
         }, error -> {
             conn = false;
             connected = false;
-            error.printStackTrace();
+            logger.log(Level.WARNING, error.getMessage(), error);
             updateLeaveGraph();
         });
 
@@ -763,9 +874,7 @@ public class Leave extends AppCompatActivity {
         waitProgress.dismiss();
         if (conn) {
             if (connected) {
-                if (balance.size() == 0 || earn.size() == 0 || shortCode.size() == 0) {
-                    // do nothing
-                } else {
+                if (!balance.isEmpty() && !earn.isEmpty() && !shortCode.isEmpty()) {
                     XAxis xAxis = chart.getXAxis();
                     xAxis.setPosition(XAxis.XAxisPosition.BOTTOM);
                     xAxis.setDrawGridLines(false);
