@@ -1,25 +1,22 @@
 package ttit.com.shuvo.ikglhrm.payRoll;
 
-import static ttit.com.shuvo.ikglhrm.Login.CompanyName;
-import static ttit.com.shuvo.ikglhrm.Login.SoftwareName;
 import static ttit.com.shuvo.ikglhrm.Login.userInfoLists;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
+
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.cardview.widget.CardView;
 
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.github.dewinjm.monthyearpicker.MonthFormat;
-import com.github.dewinjm.monthyearpicker.MonthYearPickerDialog;
 import com.github.dewinjm.monthyearpicker.MonthYearPickerDialogFragment;
 import com.github.mikephil.charting.charts.BarChart;
 import com.github.mikephil.charting.components.AxisBase;
@@ -30,18 +27,20 @@ import com.github.mikephil.charting.data.BarEntry;
 import com.github.mikephil.charting.formatter.LargeValueFormatter;
 import com.github.mikephil.charting.formatter.ValueFormatter;
 import com.github.mikephil.charting.utils.ColorTemplate;
+import com.google.android.material.card.MaterialCardView;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.sql.Connection;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import ttit.com.shuvo.ikglhrm.R;
 import ttit.com.shuvo.ikglhrm.WaitProgress;
@@ -49,14 +48,9 @@ import ttit.com.shuvo.ikglhrm.payRoll.advance.AdvanceDetails;
 import ttit.com.shuvo.ikglhrm.payRoll.paySlip.PaySlip;
 
 public class PayRollInfo extends AppCompatActivity {
-
-    TextView softName;
-    TextView compName;
-
-    Button back;
-
-    CardView paySlip;
-    CardView advance;
+    
+    MaterialCardView paySlip;
+    MaterialCardView advance;
 
     BarChart chart;
     TextView refresh;
@@ -71,44 +65,24 @@ public class PayRollInfo extends AppCompatActivity {
     ArrayList<String> salary;
 
     WaitProgress waitProgress = new WaitProgress();
-    private String message = null;
     private Boolean conn = false;
     private Boolean connected = false;
-
-    private Connection connection;
-
+    
     String emp_id = "";
     String formattedDate = "";
+
+    Logger logger = Logger.getLogger(PayRollInfo.class.getName());
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-//        getWindow().requestFeature(Window.FEATURE_NO_TITLE);
-//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-//            Window w = getWindow();
-//            //w.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
-//        }
-//        if (Build.VERSION.SDK_INT < 16) {
-//            getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-//                    WindowManager.LayoutParams.FLAG_FULLSCREEN);
-//        }
-//        View decorView = getWindow().getDecorView();
-//// Hide the status bar.
-//        int uiOptions = View.SYSTEM_UI_FLAG_FULLSCREEN;
-//        decorView.setSystemUiVisibility(uiOptions);
         setContentView(R.layout.activity_pay_roll_info);
 
         emp_id = userInfoLists.get(0).getEmp_id();
 
         paySlip = findViewById(R.id.pay_slip);
         advance = findViewById(R.id.advance_details);
-        back = findViewById(R.id.pay_roll_back);
         refresh = findViewById(R.id.refresh_graph);
-        softName = findViewById(R.id.name_of_soft_pay_roll);
-        compName = findViewById(R.id.name_of_company_pay_roll);
-        String soft = CompanyName;
-        softName.setText(soft);
-        compName.setText(SoftwareName);
 
         chart = findViewById(R.id.barchart);
 
@@ -128,16 +102,7 @@ public class PayRollInfo extends AppCompatActivity {
 
         NoOfEmp = new ArrayList<>();
 
-
-
         year = new ArrayList<>();
-
-//        year.add("MAR");
-//        year.add("APR");
-//        year.add("MAY");
-//        year.add("JUN");
-//        year.add("JUL");
-//        year.add("AUG");
 
         XAxis xAxis = chart.getXAxis();
         xAxis.setPosition(XAxis.XAxisPosition.BOTTOM);
@@ -155,20 +120,19 @@ public class PayRollInfo extends AppCompatActivity {
         chart.setDoubleTapToZoomEnabled(false);
 
         chart.getAxisRight().setEnabled(false);
-        chart.getAxisLeft().setAxisMinValue(0);
+        chart.getAxisLeft().setAxisMinimum(0);
         chart.getLegend().setEnabled(false);
 
-
-
+        
+        
         Date c = Calendar.getInstance().getTime();
-
-
 
         SimpleDateFormat df = new SimpleDateFormat("dd-MMM-yy", Locale.ENGLISH);
         SimpleDateFormat mom = new SimpleDateFormat("MMMM",Locale.ENGLISH);
-        String monnn = mom.format(c);
+        String monnn = mom.format(c).toUpperCase();
 
-        refresh.setText("Month: "+monnn);
+        String rt = "Month: "+monnn;
+        refresh.setText(rt);
 
         formattedDate = df.format(c);
 
@@ -215,225 +179,220 @@ public class PayRollInfo extends AppCompatActivity {
         getSalaryGraph();
 
 
-        paySlip.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(PayRollInfo.this, PaySlip.class);
-                startActivity(intent);
-            }
+        paySlip.setOnClickListener(v -> {
+            Intent intent = new Intent(PayRollInfo.this, PaySlip.class);
+            startActivity(intent);
         });
 
-        back.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
+        advance.setOnClickListener(v -> {
+            Intent intent = new Intent(PayRollInfo.this, AdvanceDetails.class);
+            startActivity(intent);
         });
 
-        advance.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(PayRollInfo.this, AdvanceDetails.class);
-                startActivity(intent);
+        refresh.setOnClickListener(v -> {
+
+            Date c1 = Calendar.getInstance().getTime();
+
+            String formattedYear;
+            String monthValue;
+            String lastformattedYear;
+            String lastdateView;
+
+            SimpleDateFormat df1 = new SimpleDateFormat("yyyy", Locale.ENGLISH);
+            SimpleDateFormat sdf = new SimpleDateFormat("MM",Locale.ENGLISH);
+
+            formattedYear = df1.format(c1);
+            monthValue = sdf.format(c1);
+            int nowMonNumb = Integer.parseInt(monthValue);
+            nowMonNumb = nowMonNumb - 1;
+            int lastMonNumb = nowMonNumb - 5;
+
+            if (lastMonNumb < 0) {
+                lastMonNumb = lastMonNumb + 12;
+                int formatY = Integer.parseInt(formattedYear);
+                formatY = formatY - 1;
+                lastformattedYear = String.valueOf(formatY);
+            } else {
+                lastformattedYear = formattedYear;
             }
-        });
 
-        refresh.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
+            Date today = new Date();
 
-                Date c = Calendar.getInstance().getTime();
+            Calendar calendar1 = Calendar.getInstance();
+            calendar1.setTime(today);
 
-                String formattedYear = "";
-                String monthValue = "";
-                String lastformattedYear = "";
-                String lastdateView = "";
+            calendar1.add(Calendar.MONTH, 1);
+            calendar1.set(Calendar.DAY_OF_MONTH, 1);
+            calendar1.add(Calendar.DATE, -1);
 
-                SimpleDateFormat df = new SimpleDateFormat("yyyy", Locale.ENGLISH);
-                SimpleDateFormat sdf = new SimpleDateFormat("MM",Locale.ENGLISH);
+            Date lastDayOfMonth = calendar1.getTime();
 
-                formattedYear = df.format(c);
-                monthValue = sdf.format(c);
-                int nowMonNumb = Integer.parseInt(monthValue);
-                nowMonNumb = nowMonNumb - 1;
-                int lastMonNumb = nowMonNumb - 5;
+            SimpleDateFormat sdff = new SimpleDateFormat("dd",Locale.ENGLISH);
+            lastdateView = sdff.format(lastDayOfMonth);
 
-                if (lastMonNumb < 0) {
-                    lastMonNumb = lastMonNumb + 12;
-                    int formatY = Integer.parseInt(formattedYear);
-                    formatY = formatY - 1;
-                    lastformattedYear = String.valueOf(formatY);
-                } else {
-                    lastformattedYear = formattedYear;
+            int yearSelected;
+            int monthSelected;
+            MonthFormat monthFormat = MonthFormat.LONG;
+            String customTitle = "Select Month";
+// Use the calendar for create ranges
+            Calendar calendar = Calendar.getInstance();
+            if (!formattedDate.isEmpty()) {
+                SimpleDateFormat myf = new SimpleDateFormat("dd-MMM-yy", Locale.ENGLISH);
+                Date md = null;
+                try {
+                    md = myf.parse(formattedDate);
+                } catch (ParseException e) {
+                    logger.log(Level.WARNING, e.getMessage(), e);
                 }
 
-                Date today = new Date();
+                if (md != null) {
+                    calendar.setTime(md);
+                }
+            }
+            yearSelected = calendar.get(Calendar.YEAR);
+            monthSelected = calendar.get(Calendar.MONTH);
+            calendar.clear();
+            calendar.set(Integer.parseInt(lastformattedYear), lastMonNumb, 1); // Set minimum date to show in dialog
+            long minDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
 
-                Calendar calendar1 = Calendar.getInstance();
-                calendar1.setTime(today);
-
-                calendar1.add(Calendar.MONTH, 1);
-                calendar1.set(Calendar.DAY_OF_MONTH, 1);
-                calendar1.add(Calendar.DATE, -1);
-
-                Date lastDayOfMonth = calendar1.getTime();
-
-                SimpleDateFormat sdff = new SimpleDateFormat("dd",Locale.ENGLISH);
-                lastdateView = sdff.format(lastDayOfMonth);
-
-                int yearSelected;
-                int monthSelected;
-                MonthFormat monthFormat = MonthFormat.LONG;
-                String customTitle = "Select Month";
-// Use the calendar for create ranges
-                Calendar calendar = Calendar.getInstance();
-                yearSelected = calendar.get(Calendar.YEAR);
-                monthSelected = calendar.get(Calendar.MONTH);
-                calendar.clear();
-                calendar.set(Integer.parseInt(lastformattedYear), lastMonNumb, 1); // Set minimum date to show in dialog
-                long minDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
-
-                calendar.clear();
-                calendar.set(Integer.parseInt(formattedYear), nowMonNumb, Integer.parseInt(lastdateView)); // Set maximum date to show in dialog
-                long maxDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
+            calendar.clear();
+            calendar.set(Integer.parseInt(formattedYear), nowMonNumb, Integer.parseInt(lastdateView)); // Set maximum date to show in dialog
+            long maxDate = calendar.getTimeInMillis(); // Get milliseconds of the modified date
 
 // Create instance with date ranges values
-                MonthYearPickerDialogFragment dialogFragment =  MonthYearPickerDialogFragment
-                        .getInstance(monthSelected, yearSelected, minDate, maxDate, customTitle, monthFormat);
+            MonthYearPickerDialogFragment dialogFragment =  MonthYearPickerDialogFragment
+                    .getInstance(monthSelected, yearSelected, minDate, maxDate, customTitle, monthFormat);
 
 
 
-                dialogFragment.show(getSupportFragmentManager(), null);
+            dialogFragment.show(getSupportFragmentManager(), null);
 
-                dialogFragment.setOnDateSetListener(new MonthYearPickerDialog.OnDateSetListener() {
-                    @Override
-                    public void onDateSet(int year, int monthOfYear) {
-                        System.out.println(year);
-                        System.out.println(monthOfYear);
+            dialogFragment.setOnDateSetListener((year, monthOfYear) -> {
+                System.out.println(year);
+                System.out.println(monthOfYear);
 
-                        int month = monthOfYear + 1;
-                        String monthName = "";
-                        String mon = "";
-                        String yearName = "";
+                int month = monthOfYear + 1;
+                String monthName = "";
+                String mon = "";
+                String yearName;
 
-                        if (month == 1) {
-                            monthName = "JANUARY";
-                            mon = "JAN";
-                        } else if (month == 2) {
-                            monthName = "FEBRUARY";
-                            mon = "FEB";
-                        } else if (month == 3) {
-                            monthName = "MARCH";
-                            mon = "MAR";
-                        } else if (month == 4) {
-                            monthName = "APRIL";
-                            mon = "APR";
-                        } else if (month == 5) {
-                            monthName = "MAY";
-                            mon = "MAY";
-                        } else if (month == 6) {
-                            monthName = "JUNE";
-                            mon = "JUN";
-                        } else if (month == 7) {
-                            monthName = "JULY";
-                            mon = "JUL";
-                        } else if (month == 8) {
-                            monthName = "AUGUST";
-                            mon = "AUG";
-                        } else if (month == 9) {
-                            monthName = "SEPTEMBER";
-                            mon = "SEP";
-                        } else if (month == 10) {
-                            monthName = "OCTOBER";
-                            mon = "OCT";
-                        } else if (month == 11) {
-                            monthName = "NOVEMBER";
-                            mon = "NOV";
-                        } else if (month == 12) {
-                            monthName = "DECEMBER";
-                            mon = "DEC";
-                        }
+                if (month == 1) {
+                    monthName = "JANUARY";
+                    mon = "JAN";
+                } else if (month == 2) {
+                    monthName = "FEBRUARY";
+                    mon = "FEB";
+                } else if (month == 3) {
+                    monthName = "MARCH";
+                    mon = "MAR";
+                } else if (month == 4) {
+                    monthName = "APRIL";
+                    mon = "APR";
+                } else if (month == 5) {
+                    monthName = "MAY";
+                    mon = "MAY";
+                } else if (month == 6) {
+                    monthName = "JUNE";
+                    mon = "JUN";
+                } else if (month == 7) {
+                    monthName = "JULY";
+                    mon = "JUL";
+                } else if (month == 8) {
+                    monthName = "AUGUST";
+                    mon = "AUG";
+                } else if (month == 9) {
+                    monthName = "SEPTEMBER";
+                    mon = "SEP";
+                } else if (month == 10) {
+                    monthName = "OCTOBER";
+                    mon = "OCT";
+                } else if (month == 11) {
+                    monthName = "NOVEMBER";
+                    mon = "NOV";
+                } else if (month == 12) {
+                    monthName = "DECEMBER";
+                    mon = "DEC";
+                }
 
-                        yearName  = String.valueOf(year);
-                        yearName = yearName.substring(yearName.length()-2);
+                yearName  = String.valueOf(year);
+                yearName = yearName.substring(yearName.length()-2);
 
 
-                        formattedDate = "15-"+mon+"-"+yearName;
-                        //selected_date = "01-"+mon+"-"+yearName;
-                        refresh.setText("Month: " +monthName);
-                        SimpleDateFormat sss = new SimpleDateFormat("dd-MMM-yyyy", Locale.ENGLISH);
+                formattedDate = "15-"+mon+"-"+yearName;
+                //selected_date = "01-"+mon+"-"+yearName;
+                String rtt = "Month: " +monthName;
+                refresh.setText(rtt);
+                SimpleDateFormat sss = new SimpleDateFormat("dd-MMM-yyyy", Locale.ENGLISH);
 
-                        Date today = null;
-                        try {
-                            today = sss.parse(formattedDate);
-                        } catch (ParseException e) {
-                            e.printStackTrace();
-                        }
+                Date today1 = null;
+                try {
+                    today1 = sss.parse(formattedDate);
+                } catch (ParseException e) {
+                    logger.log(Level.WARNING, e.getMessage(), e);
+                }
 
-                        Calendar calendar1 = Calendar.getInstance();
-                        if (today != null) {
-                            calendar1.setTime(today);
-                            calendar1.add(Calendar.MONTH, 1);
-                            calendar1.set(Calendar.DAY_OF_MONTH, 1);
-                            calendar1.add(Calendar.DATE, -1);
+                Calendar calendar11 = Calendar.getInstance();
+                if (today1 != null) {
+                    calendar11.setTime(today1);
+                    calendar11.add(Calendar.MONTH, 1);
+                    calendar11.set(Calendar.DAY_OF_MONTH, 1);
+                    calendar11.add(Calendar.DATE, -1);
 
-                            Date lastDayOfMonth = calendar1.getTime();
+                    Date lastDayOfMonth1 = calendar11.getTime();
 
-                            SimpleDateFormat sdff = new SimpleDateFormat("dd",Locale.ENGLISH);
-                            String llll = sdff.format(lastDayOfMonth);
-                            formattedDate =  llll+ "-" + mon +"-"+ yearName;
+                    SimpleDateFormat sdff1 = new SimpleDateFormat("dd",Locale.ENGLISH);
+                    String llll = sdff1.format(lastDayOfMonth1);
+                    formattedDate =  llll+ "-" + mon +"-"+ yearName;
 
-                            months = new ArrayList<>();
+                    months = new ArrayList<>();
 
-                            calendar1.add(Calendar.MONTH, -1);
-                            String previousMonthYear  = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar1.getTime());
-                            previousMonthYear = previousMonthYear.toUpperCase();
-                            months.add(new SalaryMonthList(previousMonthYear,"0"));
-                            System.out.println(previousMonthYear);
+                    calendar11.add(Calendar.MONTH, -1);
+                    String previousMonthYear1 = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar11.getTime());
+                    previousMonthYear1 = previousMonthYear1.toUpperCase();
+                    months.add(new SalaryMonthList(previousMonthYear1,"0"));
+                    System.out.println(previousMonthYear1);
 
-                            calendar1.add(Calendar.MONTH, -1);
-                            previousMonthYear  = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar1.getTime());
-                            previousMonthYear = previousMonthYear.toUpperCase();
-                            months.add(new SalaryMonthList(previousMonthYear,"0"));
-                            System.out.println(previousMonthYear);
+                    calendar11.add(Calendar.MONTH, -1);
+                    previousMonthYear1 = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar11.getTime());
+                    previousMonthYear1 = previousMonthYear1.toUpperCase();
+                    months.add(new SalaryMonthList(previousMonthYear1,"0"));
+                    System.out.println(previousMonthYear1);
 
-                            calendar1.add(Calendar.MONTH, -1);
-                            previousMonthYear  = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar1.getTime());
-                            previousMonthYear = previousMonthYear.toUpperCase();
-                            months.add(new SalaryMonthList(previousMonthYear,"0"));
-                            System.out.println(previousMonthYear);
+                    calendar11.add(Calendar.MONTH, -1);
+                    previousMonthYear1 = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar11.getTime());
+                    previousMonthYear1 = previousMonthYear1.toUpperCase();
+                    months.add(new SalaryMonthList(previousMonthYear1,"0"));
+                    System.out.println(previousMonthYear1);
 
-                            calendar1.add(Calendar.MONTH, -1);
-                            previousMonthYear  = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar1.getTime());
-                            previousMonthYear = previousMonthYear.toUpperCase();
-                            months.add(new SalaryMonthList(previousMonthYear,"0"));
-                            System.out.println(previousMonthYear);
+                    calendar11.add(Calendar.MONTH, -1);
+                    previousMonthYear1 = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar11.getTime());
+                    previousMonthYear1 = previousMonthYear1.toUpperCase();
+                    months.add(new SalaryMonthList(previousMonthYear1,"0"));
+                    System.out.println(previousMonthYear1);
 
-                            calendar1.add(Calendar.MONTH, -1);
-                            previousMonthYear  = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar1.getTime());
-                            previousMonthYear = previousMonthYear.toUpperCase();
-                            months.add(new SalaryMonthList(previousMonthYear,"0"));
-                            System.out.println(previousMonthYear);
+                    calendar11.add(Calendar.MONTH, -1);
+                    previousMonthYear1 = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar11.getTime());
+                    previousMonthYear1 = previousMonthYear1.toUpperCase();
+                    months.add(new SalaryMonthList(previousMonthYear1,"0"));
+                    System.out.println(previousMonthYear1);
 
-                            calendar1.add(Calendar.MONTH, -1);
-                            previousMonthYear  = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar1.getTime());
-                            previousMonthYear = previousMonthYear.toUpperCase();
-                            months.add(new SalaryMonthList(previousMonthYear,"0"));
-                            System.out.println(previousMonthYear);
+                    calendar11.add(Calendar.MONTH, -1);
+                    previousMonthYear1 = new SimpleDateFormat("MMM-yy", Locale.ENGLISH).format(calendar11.getTime());
+                    previousMonthYear1 = previousMonthYear1.toUpperCase();
+                    months.add(new SalaryMonthList(previousMonthYear1,"0"));
+                    System.out.println(previousMonthYear1);
 
 
-                        }
+                }
 
 
 
 //                        new Check().execute();
-                        getSalaryGraph();
-                        chart.resetZoom();
-                        chart.fitScreen();
-                    }
-                });
+                getSalaryGraph();
+                chart.resetZoom();
+                chart.fitScreen();
+            });
 
-            }
         });
 
 
@@ -473,7 +432,7 @@ public class PayRollInfo extends AppCompatActivity {
 //            int     exitValue = ipProcess.waitFor();
 //            return (exitValue == 0);
 //        }
-//        catch (IOException | InterruptedException e)          { e.printStackTrace(); }
+//        catch (IOException | InterruptedException e)          { logger.log(Level.WARNING, e.getMessage(), e); }
 //
 //        return false;
 //    }
@@ -683,7 +642,7 @@ public class PayRollInfo extends AppCompatActivity {
 //
 //            //   Toast.makeText(MainActivity.this, ""+e,Toast.LENGTH_LONG).show();
 //            Log.i("ERRRRR", e.getLocalizedMessage());
-//            e.printStackTrace();
+//            logger.log(Level.WARNING, e.getMessage(), e);
 //        }
 //    }
 
@@ -695,9 +654,16 @@ public class PayRollInfo extends AppCompatActivity {
 
         salaryMonthLists = new ArrayList<>();
 
+        BarData salaryData = new BarData();
+        chart.setData(salaryData);
+        chart.getData().clearValues();
+        chart.notifyDataSetChanged();
+        chart.clear();
+        chart.invalidate();
+
         RequestQueue requestQueue = Volley.newRequestQueue(PayRollInfo.this);
 
-        String salaryDataUrl = "http://103.56.208.123:8001/apex/ttrams/dashboard/getSalaryAndMonth/"+emp_id+"/"+formattedDate+"";
+        String salaryDataUrl = api_url_front + "dashboard/getSalaryAndMonth/"+emp_id+"/"+formattedDate;
 
         StringRequest salaryMonthReq = new StringRequest(Request.Method.GET, salaryDataUrl, response -> {
             conn = true;
@@ -719,13 +685,13 @@ public class PayRollInfo extends AppCompatActivity {
             }
             catch (JSONException e) {
                 connected = false;
-                e.printStackTrace();
+                logger.log(Level.WARNING, e.getMessage(), e);
                 updateSalaryGraph();
             }
         }, error -> {
             conn = false;
             connected = false;
-            error.printStackTrace();
+            logger.log(Level.WARNING, error.getMessage(), error);
             updateSalaryGraph();
         });
 
@@ -743,7 +709,7 @@ public class PayRollInfo extends AppCompatActivity {
                 for (int i = 0; i < salaryMonthLists.size(); i++) {
                     for (int j = 0; j < months.size(); j++) {
                         String month = months.get(j).getMonth();
-                        month = month.substring(0, month.length() -3);
+//                        month = month.substring(0, month.length() -3);
                         if (month.equals(salaryMonthLists.get(i).getMonth())) {
                             months.get(j).setSalary(salaryMonthLists.get(i).getSalary());
                         }
@@ -751,7 +717,6 @@ public class PayRollInfo extends AppCompatActivity {
                 }
 
                 for (int i = months.size()-1; i >= 0; i--) {
-
                     monthName.add(months.get(i).getMonth());
                     salary.add(months.get(i).getSalary());
 
@@ -769,12 +734,24 @@ public class PayRollInfo extends AppCompatActivity {
                 chart.animateY(1000);
 
                 BarData data1 = new BarData(bardataset);
-                bardataset.setColors(ColorTemplate.VORDIPLOM_COLORS);
+                final int[] barColors = new int[]{
+                        Color.rgb(146, 197, 249),
+                        Color.rgb(154, 216, 216),
+                        Color.rgb(175, 220, 143),
+                        Color.rgb(251, 224, 114),
+                        Color.rgb(113, 164, 201),
+                        Color.rgb(188, 228, 216),
+
+                        Color.rgb(129, 236, 236),
+                        Color.rgb(255, 118, 117),
+                        Color.rgb(253, 121, 168),
+                        Color.rgb(96, 163, 188)};
+                bardataset.setColors(ColorTemplate.createColors(barColors));
 
                 bardataset.setBarBorderColor(Color.DKGRAY);
                 bardataset.setValueTextSize(11);
+                bardataset.setValueFormatter(new LargeValueFormatter());
                 chart.setData(data1);
-
                 chart.getXAxis().setValueFormatter(new MyAxisValueFormatter(monthName));
                 chart.getAxisLeft().setValueFormatter(new LargeValueFormatter());
             }
