@@ -1,12 +1,14 @@
 package ttit.com.shuvo.ikglhrm.attendance.att_widget;
 
 
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.CENTER_API_FRONT;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
+
 import android.Manifest;
 import android.annotation.SuppressLint;
 import android.appwidget.AppWidgetManager;
 import android.content.ComponentName;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.location.Address;
@@ -16,22 +18,17 @@ import android.location.LocationManager;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.AsyncTask;
-import android.os.Looper;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
 import android.widget.RemoteViews;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AlertDialog;
 import androidx.core.app.ActivityCompat;
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 import androidx.work.Worker;
 import androidx.work.WorkerParameters;
 
-import com.android.volley.AuthFailureError;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
@@ -43,9 +40,6 @@ import com.google.android.gms.location.LocationResult;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.location.Priority;
 import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.Task;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -54,6 +48,7 @@ import org.json.JSONObject;
 import java.io.IOException;
 import java.sql.Timestamp;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.HashMap;
@@ -62,8 +57,7 @@ import java.util.Locale;
 import java.util.Map;
 
 import ttit.com.shuvo.ikglhrm.R;
-import ttit.com.shuvo.ikglhrm.attendance.giveAttendance.AttendanceGive;
-import ttit.com.shuvo.ikglhrm.attendance.trackService.Service;
+import ttit.com.shuvo.ikglhrm.attendance.giveAttendance.arraylists.AreaList;
 
 public class WidgetAttendanceWorker extends Worker {
 
@@ -82,11 +76,11 @@ public class WidgetAttendanceWorker extends Worker {
     String timeToShow = "";
     private Boolean conn = false;
     private Boolean connected = false;
-    String officeLatitude = "";
-    String officeLongitude = "";
-    String coverage = "";
+
     String machineCode = "";
     private LocationCallback locationCallback;
+    ArrayList<AreaList> areaLists;
+    SharedPreferences sharedPreferences;
 
     public WidgetAttendanceWorker(@NonNull Context context, @NonNull WorkerParameters workerParams) {
         super(context, workerParams);
@@ -97,6 +91,7 @@ public class WidgetAttendanceWorker extends Worker {
     @Override
     public Result doWork() {
         emp_id = getInputData().getString(WORKER_W_EMP_ID);
+        api_url_front = sharedPreferences.getString(CENTER_API_FRONT,null);
 
         fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(getApplicationContext());
         locationManager = (LocationManager) mContext.getSystemService(Context.LOCATION_SERVICE);
@@ -113,81 +108,84 @@ public class WidgetAttendanceWorker extends Worker {
 
         boolean gps = locationManager.isProviderEnabled(LocationManager.GPS_PROVIDER);
 
-        if (gps) {
-            RemoteViews views = new RemoteViews(mContext.getPackageName(), R.layout.attendance_widget);
-            views.setViewVisibility(R.id.attendance_widget_button, View.GONE);
-            views.setViewVisibility(R.id.attendance_widget_progress_bar, View.VISIBLE);
+        if (api_url_front != null) {
+            if (!api_url_front.isEmpty()) {
+                if (gps) {
+                    RemoteViews views = new RemoteViews(mContext.getPackageName(), R.layout.attendance_widget);
+                    views.setViewVisibility(R.id.attendance_widget_button, View.GONE);
+                    views.setViewVisibility(R.id.attendance_widget_progress_bar, View.VISIBLE);
 
-            AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(mContext);
-            appWidgetManager.updateAppWidget(new ComponentName(mContext, AttendanceWidget.class), views);
-            locationCallback = new LocationCallback() {
-                @Override
-                public void onLocationResult(LocationResult locationResult) {
-                    super.onLocationResult(locationResult);
-                }
-            };
+                    AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(mContext);
+                    appWidgetManager.updateAppWidget(new ComponentName(mContext, AttendanceWidget.class), views);
+                    locationCallback = new LocationCallback() {
+                        @Override
+                        public void onLocationResult(@NonNull LocationResult locationResult) {
+                            super.onLocationResult(locationResult);
+                        }
+                    };
 //            LocationRequest mLocationRequest = LocationRequest.create();
 //            mLocationRequest.setInterval(5000);
 //            mLocationRequest.setFastestInterval(1000);
 //            mLocationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
 
-            try {
-                fusedLocationProviderClient.requestLocationUpdates(locationRequest, null);
-            } catch (SecurityException unlikely) {
-                //Utils.setRequestingLocationUpdates(this, false);
-                Log.e("TAG", "Lost location permission. Could not request updates. " + unlikely);
-                updateWidget();
-                getNotification("Attendance System", "Lost location permission. Could not request updates.");
-            }
-            try {
-                fusedLocationProviderClient
-                        .getLastLocation()
-                        .addOnCompleteListener(task -> {
-                            if (task.isSuccessful() && task.getResult() != null) {
-                                Location mLocation = task.getResult();
-                                Log.d("TAG", "Location : " + mLocation);
-                                fusedLocationProviderClient.removeLocationUpdates(locationCallback);
-                                SimpleDateFormat df = new SimpleDateFormat("dd-MMM-yy, hh:mm:ss aa", Locale.getDefault());
-                                SimpleDateFormat dftoShow = new SimpleDateFormat("hh:mm aa", Locale.getDefault());
-                                lastLatLongitude[0] = new LatLng(mLocation.getLatitude(), mLocation.getLongitude());
-                                lat = String.valueOf(lastLatLongitude[0].latitude);
-                                lon = String.valueOf(lastLatLongitude[0].longitude);
-                                Date c = Calendar.getInstance().getTime();
-                                Date date = new Date();
-                                ts = new Timestamp(date.getTime());
-                                System.out.println(ts.toString());
-                                inTime = df.format(c);
-                                timeToShow = dftoShow.format(c);
-                                System.out.println("IN TIME : " + inTime);
-                                getOfficeLocation();
-                            } else {
-                                task.addOnFailureListener(new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        e.printStackTrace();
+                    try {
+                        fusedLocationProviderClient.requestLocationUpdates(locationRequest, null);
+                    } catch (SecurityException unlikely) {
+                        //Utils.setRequestingLocationUpdates(this, false);
+                        Log.e("TAG", "Lost location permission. Could not request updates. " + unlikely);
+                        updateWidget();
+                        getNotification("Attendance System", "Lost location permission. Could not request updates.");
+                    }
+                    try {
+                        fusedLocationProviderClient
+                                .getLastLocation()
+                                .addOnCompleteListener(task -> {
+                                    if (task.isSuccessful() && task.getResult() != null) {
+                                        Location mLocation = task.getResult();
+                                        Log.d("TAG", "Location : " + mLocation);
+                                        fusedLocationProviderClient.removeLocationUpdates(locationCallback);
+                                        SimpleDateFormat df = new SimpleDateFormat("dd-MMM-yy, hh:mm:ss aa", Locale.getDefault());
+                                        SimpleDateFormat dftoShow = new SimpleDateFormat("hh:mm aa", Locale.getDefault());
+                                        lastLatLongitude[0] = new LatLng(mLocation.getLatitude(), mLocation.getLongitude());
+                                        lat = String.valueOf(lastLatLongitude[0].latitude);
+                                        lon = String.valueOf(lastLatLongitude[0].longitude);
+                                        Date c = Calendar.getInstance().getTime();
+                                        Date date = new Date();
+                                        ts = new Timestamp(date.getTime());
+                                        System.out.println(ts);
+                                        inTime = df.format(c);
+                                        timeToShow = dftoShow.format(c);
+                                        System.out.println("IN TIME : " + inTime);
+                                        getOfficeLocation();
+                                    } else {
+                                        task.addOnFailureListener(e -> {
+                                        });
+                                        Log.w("TAG", "Failed to get location.");
+                                        updateWidget();
+                                        getNotification("Attendance System", "Failed to get location.");
                                     }
+                                })
+                                .addOnFailureListener(e -> {
                                 });
-                                Log.w("TAG", "Failed to get location.");
-                                updateWidget();
-                                getNotification("Attendance System", "Failed to get location.");
-                            }
-                        })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                e.printStackTrace();
-                            }
-                        });
-            }
-            catch (SecurityException unlikely) {
-                Log.e("TAG", "Lost location permission." + unlikely);
-                updateWidget();
-                getNotification("Attendance System", "Lost location permission.");
-            }
+                    }
+                    catch (SecurityException unlikely) {
+                        Log.e("TAG", "Lost location permission." + unlikely);
+                        updateWidget();
+                        getNotification("Attendance System", "Lost location permission.");
+                    }
 
-        } else {
-            getNotification("Attendance System", "Your GPS is disabled. Please enable it and try again.");
+                } else {
+                    getNotification("Attendance System", "Your GPS is disabled. Please enable it and try again.");
+                }
+            }
+            else {
+                getNotification("Attendance System", "Could not find destined server to upload attendance time.");
+            }
         }
+        else {
+            getNotification("Attendance System", "Could not find destined server to upload attendance time.");
+        }
+
 //        getOfficeLocation();
 
         return Result.success();
@@ -226,8 +224,9 @@ public class WidgetAttendanceWorker extends Worker {
         getOfficeLocation();
     }
     public void getOfficeLocation() {
+        areaLists = new ArrayList<>();
 
-        String offLocationUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/getOffLatLong/"+emp_id+"";
+        String offLocationUrl = api_url_front + "attendance/getNewOffLatLong?emp_id="+emp_id;
 
         RequestQueue requestQueue = Volley.newRequestQueue(mContext);
 
@@ -241,21 +240,30 @@ public class WidgetAttendanceWorker extends Worker {
                     JSONArray array = new JSONArray(items);
                     for (int i = 0; i < array.length(); i++) {
                         JSONObject offLocInfo = array.getJSONObject(i);
-                        officeLatitude = offLocInfo.getString("coa_latitude").equals("null") ? null : offLocInfo.getString("coa_latitude");
-                        officeLongitude = offLocInfo.getString("coa_longitude").equals("null") ? null : offLocInfo.getString("coa_longitude");
-                        coverage = offLocInfo.getString("coa_coverage").equals("null") ? null : offLocInfo.getString("coa_coverage");
+                        String coa_latitude = offLocInfo.getString("coa_latitude")
+                                .equals("null") ? null : offLocInfo.getString("coa_latitude");
+                        String coa_longitude = offLocInfo.getString("coa_longitude")
+                                .equals("null") ? null : offLocInfo.getString("coa_longitude");
+                        String coa_coverage = offLocInfo.getString("coa_coverage")
+                                .equals("null") ? null : offLocInfo.getString("coa_coverage");
+                        String co_id = offLocInfo.getString("coa_id")
+                                .equals("null") ? null : offLocInfo.getString("coa_id");
+                        String code = offLocInfo.getString("machine_code")
+                                .equals("null") ? null : offLocInfo.getString("machine_code");
+                        String can_give = offLocInfo.getString("can_give")
+                                .equals("null") ? "0" : offLocInfo.getString("can_give");
+
+                        areaLists.add(new AreaList(coa_latitude,coa_longitude,coa_coverage,co_id,code,can_give.equals("1")));
                     }
                 }
                 connected = true;
                 updateInfo();
             }
             catch (JSONException e) {
-                e.printStackTrace();
                 connected = false;
                 updateInfo();
             }
         }, error -> {
-            error.printStackTrace();
             conn = false;
             connected = false;
             updateInfo();
@@ -270,40 +278,83 @@ public class WidgetAttendanceWorker extends Worker {
                 conn = false;
                 connected = false;
                 if (!inTime.isEmpty()) {
-                    LatLng c_latLng = new LatLng(0,0);
-                    if (officeLatitude != null && officeLongitude != null) {
-                        if (!officeLatitude.isEmpty() && !officeLongitude.isEmpty()) {
-                            c_latLng = new LatLng(Double.parseDouble(officeLatitude),Double.parseDouble(officeLongitude));
-                        }
-                    }
-                    if (c_latLng.latitude != 0 && c_latLng.longitude != 0) {
+                    if (!areaLists.isEmpty()) {
+                        LatLng c_latLng = new LatLng(0,0);
                         float[] distance = new float[1];
-                        Location.distanceBetween(c_latLng.latitude,c_latLng.longitude,lastLatLongitude[0].latitude,lastLatLongitude[0].longitude,distance);
+                        boolean found = false;
+                        float prev_distance = 0;
+                        String prev_mach_code = "";
 
-                        float radius = 0;
-                        if (coverage != null) {
-                            if (!coverage.isEmpty()) {
-                                radius = Float.parseFloat(coverage);
+                        machineCode = "";
+
+                        for (int i = 0; i < areaLists.size(); i++) {
+                            String officeLatitude = areaLists.get(i).getLatitude();
+                            String officeLongitude = areaLists.get(i).getLongitude();
+                            String coverage = areaLists.get(i).getCoverage();
+
+                            if (officeLatitude != null && officeLongitude != null) {
+                                if (!officeLatitude.isEmpty() && !officeLongitude.isEmpty()) {
+                                    c_latLng = new LatLng(Double.parseDouble(officeLatitude),Double.parseDouble(officeLongitude));
+                                }
+                            }
+
+                            if (c_latLng.latitude != 0 && c_latLng.longitude != 0) {
+                                Location.distanceBetween(c_latLng.latitude,c_latLng.longitude,lastLatLongitude[0].latitude,lastLatLongitude[0].longitude,distance);
+
+                                float radius = 0;
+                                if (coverage != null) {
+                                    if (!coverage.isEmpty()) {
+                                        radius = Float.parseFloat(coverage);
+                                    }
+                                }
+
+                                machineCode = areaLists.get(i).getMachine_code();
+
+                                if (distance[0] <= radius) {
+                                    found = true;
+                                    prev_mach_code = machineCode;
+                                    break;
+                                }
+                                else {
+                                    float dd = distance[0] - radius;
+                                    if (prev_distance == 0) {
+                                        prev_distance = dd;
+                                        prev_mach_code = machineCode;
+                                    }
+                                    else if (dd < prev_distance) {
+                                        prev_distance = dd;
+                                        prev_mach_code = machineCode;
+                                    }
+                                }
                             }
                         }
 
-                        if (radius == 0) {
-                            machineCode = "3";
-                        } else {
-                            machineCode = "1";
-                        }
+                        machineCode = prev_mach_code;
 
-                        if (distance[0] <= radius || radius == 0) {
+                        if (found) {
                             new CheckAddress().execute();
                         }
                         else {
-                            updateWidget();
-                            getNotification("Attendance System","You are not around your office area");
+                            if (areaLists.get(0).isCanGive()) {
+                                if (machineCode.isEmpty()) {
+                                    machineCode = areaLists.get(0).getMachine_code();
+                                }
+                                new CheckAddress().execute();
+                            }
+                            else {
+                                updateWidget();
+                                if (prev_distance == 0) {
+                                    getNotification("Attendance System", "You are not around your office area");
+                                }
+                                else {
+                                    getNotification("Attendance System", "You are not around your office area. You are "+prev_distance+" Meters away from office.");
+                                }
+                            }
                         }
                     }
                     else {
-                        machineCode = "3";
-                        new CheckAddress().execute();
+                        updateWidget();
+                        getNotification("Attendance System", "You don't have any permission to give attendance from this app. Please contact with administrator");
                     }
                 }
                 else {
@@ -321,6 +372,7 @@ public class WidgetAttendanceWorker extends Worker {
             getNotification("Attendance System","Please Check Your Internet Connection.");
         }
     }
+
     public boolean isConnected() {
         boolean connected = false;
         boolean isMobile = false;
@@ -343,7 +395,7 @@ public class WidgetAttendanceWorker extends Worker {
             int     exitValue = ipProcess.waitFor();
             return (exitValue == 0);
         }
-        catch (IOException | InterruptedException e)          { e.printStackTrace(); }
+        catch (IOException | InterruptedException ignored) {}
 
         return false;
     }
@@ -389,15 +441,12 @@ public class WidgetAttendanceWorker extends Worker {
             List<Address> addresses = geocoder.getFromLocation(lat, lng, 1);
             if (Geocoder.isPresent()) {
                 Address obj = addresses.get(0);
-                String adds = obj.getAddressLine(0);
-                address = adds;
+                address = obj.getAddressLine(0);
                 System.out.println("Ekhane ashbe 1st");
             } else {
                 address = "";
             }
-
         } catch (IOException e) {
-            e.printStackTrace();
             address = "";
         }
     }
@@ -406,7 +455,7 @@ public class WidgetAttendanceWorker extends Worker {
         conn = false;
         connected = false;
 
-        String attendaceUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/giveAttendance";
+        String attendaceUrl = api_url_front + "attendance/giveAttendance";
 
         RequestQueue requestQueue = Volley.newRequestQueue(mContext);
 
@@ -425,18 +474,16 @@ public class WidgetAttendanceWorker extends Worker {
                 updateLayout();
             }
             catch (JSONException e) {
-                e.printStackTrace();
                 connected = false;
                 updateLayout();
             }
         }, error -> {
-            error.printStackTrace();
             conn = false;
             connected = false;
             updateLayout();
         }) {
             @Override
-            public Map<String, String> getHeaders() throws AuthFailureError {
+            public Map<String, String> getHeaders() {
                 Map<String, String> headers = new HashMap<>();
                 headers.put("P_EMP_ID",emp_id);
                 headers.put("P_PUNCH_TIME",ts.toString());
@@ -480,9 +527,10 @@ public class WidgetAttendanceWorker extends Worker {
         appWidgetManager.updateAppWidget(new ComponentName(mContext, AttendanceWidget.class),views);
     }
 
+    @SuppressLint("MissingPermission")
     public void getNotification(String title, String msg) {
         NotificationCompat.Builder builder = new NotificationCompat.Builder(mContext, mContext.getString(R.string.att_channel_id))
-                .setSmallIcon(R.drawable.thrn_logo)
+                .setSmallIcon(R.drawable.hrm_new_icon_wb)
                 .setContentTitle(title)
                 .setContentText(msg)
                 .setPriority(NotificationCompat.PRIORITY_DEFAULT);
@@ -493,13 +541,6 @@ public class WidgetAttendanceWorker extends Worker {
 
     private void startLocationUpdates() {
         if (ActivityCompat.checkSelfPermission(mContext, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(mContext, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            // TODO: Consider calling
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
             updateWidget();
             getNotification("Attendance System","Please check your Location Permission to access this feature.");
             return;
