@@ -1,9 +1,7 @@
 package ttit.com.shuvo.ikglhrm.attendance.update.dialogue;
 
 import android.app.Dialog;
-import android.content.DialogInterface;
 import android.os.Bundle;
-import android.text.InputType;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
@@ -17,6 +15,8 @@ import androidx.appcompat.app.AppCompatDialogFragment;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
 
+import java.util.Objects;
+
 import ttit.com.shuvo.ikglhrm.R;
 import ttit.com.shuvo.ikglhrm.attendance.approve.AttendanceApprove;
 import ttit.com.shuvo.ikglhrm.attendance.update.AttendanceUpdate;
@@ -26,8 +26,8 @@ import ttit.com.shuvo.ikglhrm.leaveAll.leaveApprove.LeaveApprove;
 public class DialogueText extends AppCompatDialogFragment {
 
     private TextInputEditText editText;
-    private TextInputLayout textLay;
-    private Button getAdd;
+    TextInputLayout textLay;
+    Button getAdd;
     AppCompatActivity activity;
 
 
@@ -37,8 +37,8 @@ public class DialogueText extends AppCompatDialogFragment {
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
 
 
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        LayoutInflater inflater = getActivity().getLayoutInflater();
+        AlertDialog.Builder builder = new AlertDialog.Builder(requireActivity());
+        LayoutInflater inflater = requireActivity().getLayoutInflater();
 
         View view = inflater.inflate(R.layout.textdialogue, null);
 
@@ -87,47 +87,45 @@ public class DialogueText extends AppCompatDialogFragment {
         }
 
 
-        dialog.setButton(Dialog.BUTTON_POSITIVE, "OK", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
+        dialog.setButton(Dialog.BUTTON_POSITIVE, "OK", (dialog1, which) -> {
 
-                String text = editText.getText().toString();
-                if (AttendanceUpdate.number == 1) {
-                    AttendanceUpdate.reasonDesc.setText(text);
-                    if (text.isEmpty()) {
-                        AttendanceUpdate.reasonLay.setHint("Write Reason Description:");
-                    } else {
-                        AttendanceUpdate.reasonLay.setHint("Reason Description:");
-                    }
-
-                    AttendanceUpdate.errorReasonDesc.setVisibility(View.GONE);
-                } else if (AttendanceUpdate.number == 2) {
-                    if (text.isEmpty()) {
-                        AttendanceUpdate.addStaLay.setHint("Write Address During Outside of Station:");
-                    } else {
-                        AttendanceUpdate.addStaLay.setHint("Address During Outside of Station:");
-                    }
-
-                    AttendanceUpdate.addressStation.setText(text);
-                } else if (LeaveApplication.leaveNumber == 1) {
-                    LeaveApplication.otherReason.setText(text);
-                    if (text.isEmpty()) {
-                        LeaveApplication.otherReasonInputLay.setHint("Write Your Other Reason:");
-                    } else {
-                        LeaveApplication.otherReasonInputLay.setHint("Other Reason:");
-                    }
-
-                    LeaveApplication.errorReason.setVisibility(View.GONE);
-
-                } else if (LeaveApplication.leaveNumber == 2) {
-                    if (text.isEmpty()) {
-                        LeaveApplication.leaveAddressLay.setHint("Write Address During Leave:");
-                    } else {
-                        LeaveApplication.leaveAddressLay.setHint("Address During Leave:");
-                    }
-                    LeaveApplication.leaveAddress.setText(text);
-
+            String text = Objects.requireNonNull(editText.getText()).toString();
+            if (AttendanceUpdate.number == 1) {
+                AttendanceUpdate.reasonDesc.setText(text);
+                if (text.isEmpty()) {
+                    AttendanceUpdate.reasonLay.setHint("Write Reason Description:");
+                } else {
+                    AttendanceUpdate.reasonLay.setHint("Reason Description:");
                 }
+
+                AttendanceUpdate.errorReasonDesc.setVisibility(View.GONE);
+            } else if (AttendanceUpdate.number == 2) {
+                if (text.isEmpty()) {
+                    AttendanceUpdate.addStaLay.setHint("Write Address During Outside of Station:");
+                } else {
+                    AttendanceUpdate.addStaLay.setHint("Address During Outside of Station:");
+                }
+
+                AttendanceUpdate.addressStation.setText(text);
+            } else if (LeaveApplication.leaveNumber == 1) {
+                LeaveApplication.otherReason.setText(text);
+                if (text.isEmpty()) {
+                    LeaveApplication.otherReasonInputLay.setHint("Write Your Other Reason:");
+                } else {
+                    LeaveApplication.otherReasonInputLay.setHint("Other Reason:");
+                }
+
+                LeaveApplication.errorReason.setVisibility(View.GONE);
+
+            } else if (LeaveApplication.leaveNumber == 2) {
+                if (text.isEmpty()) {
+                    LeaveApplication.leaveAddressLay.setHint("Write Address During Leave:");
+                } else {
+                    LeaveApplication.leaveAddressLay.setHint("Address During Leave:");
+                }
+                LeaveApplication.leaveAddress.setText(text);
+
+            }
 //                else if (AttendanceReqUpdate.number_update == 1) {
 //                    AttendanceReqUpdate.reasonDescUpdate.setText(text);
 //                    AttendanceReqUpdate.errorReasonDescUpdate.setVisibility(View.GONE);
@@ -135,41 +133,37 @@ public class DialogueText extends AppCompatDialogFragment {
 //                else if (AttendanceReqUpdate.number_update == 2) {
 //                    AttendanceReqUpdate.addressStationUpdate.setText(text);
 //                }
-                else if (AttendanceApprove.number == 1) {
-                    AttendanceApprove.comments.setText(text);
-                    if (text.isEmpty()) {
-                        AttendanceApprove.commentsLay.setHint("Write Approve/Reject Comments:");
-                    } else {
-                        AttendanceApprove.commentsLay.setHint("Comments:");
-                    }
-                } else if (LeaveApprove.number == 1) {
-                    LeaveApprove.comments.setText(text);
-                    if (text.isEmpty()) {
-                        LeaveApprove.commentsLay.setHint("Write Approve/Reject Comments:");
-                    } else {
-                        LeaveApprove.commentsLay.setHint("Comments:");
-                    }
+            else if (AttendanceApprove.number == 1) {
+                AttendanceApprove.comments.setText(text);
+                if (text.isEmpty()) {
+                    AttendanceApprove.commentsLay.setHint("Write Approve/Reject Comments:");
+                } else {
+                    AttendanceApprove.commentsLay.setHint("Comments:");
                 }
-                AttendanceUpdate.number = 0;
-//                AttendanceReqUpdate.number_update = 0;
-                LeaveApplication.leaveNumber = 0;
-                AttendanceApprove.number = 0;
-                LeaveApprove.number = 0;
-
-                dialog.dismiss();
+            } else if (LeaveApprove.number == 1) {
+                LeaveApprove.comments.setText(text);
+                if (text.isEmpty()) {
+                    LeaveApprove.commentsLay.setHint("Write Approve/Reject Comments:");
+                } else {
+                    LeaveApprove.commentsLay.setHint("Comments:");
+                }
             }
+            AttendanceUpdate.number = 0;
+//                AttendanceReqUpdate.number_update = 0;
+            LeaveApplication.leaveNumber = 0;
+            AttendanceApprove.number = 0;
+            LeaveApprove.number = 0;
+
+            dialog1.dismiss();
         });
 
-        dialog.setButton(Dialog.BUTTON_NEGATIVE, "CANCEL", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                AttendanceUpdate.number = 0;
+        dialog.setButton(Dialog.BUTTON_NEGATIVE, "CANCEL", (dialog12, which) -> {
+            AttendanceUpdate.number = 0;
 //                AttendanceReqUpdate.number_update = 0;
-                LeaveApplication.leaveNumber = 0;
-                AttendanceApprove.number = 0;
-                LeaveApprove.number = 0;
-                dialog.dismiss();
-            }
+            LeaveApplication.leaveNumber = 0;
+            AttendanceApprove.number = 0;
+            LeaveApprove.number = 0;
+            dialog12.dismiss();
         });
 
         return dialog;
