@@ -2,7 +2,6 @@ package ttit.com.shuvo.ikglhrm.attendance.update.dialogue;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.content.DialogInterface;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -18,12 +17,15 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import ttit.com.shuvo.ikglhrm.R;
 import ttit.com.shuvo.ikglhrm.WaitProgress;
 
 import static ttit.com.shuvo.ikglhrm.attendance.update.AttendanceUpdate.shift_osm_id;
 import static ttit.com.shuvo.ikglhrm.attendance.update.AttendanceUpdate.showShiftNumber;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
 
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
@@ -36,8 +38,8 @@ import org.json.JSONObject;
 public class ShowShift extends AppCompatDialogFragment {
 
     private RecyclerView apptRecyclerView;
-    public static ShowShiftAdapter showShiftAdapter;
-    private RecyclerView.LayoutManager apptLayout;
+    ShowShiftAdapter showShiftAdapter;
+    RecyclerView.LayoutManager apptLayout;
 
 
     WaitProgress waitProgress = new WaitProgress();
@@ -56,6 +58,8 @@ public class ShowShift extends AppCompatDialogFragment {
 
     Context mContext;
 
+    Logger logger = Logger.getLogger(ShowShift.class.getName());
+
     public ShowShift(Context mContext) {
         this.mContext = mContext;
     }
@@ -64,8 +68,8 @@ public class ShowShift extends AppCompatDialogFragment {
     @Override
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
 
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        LayoutInflater inflater = getActivity().getLayoutInflater();
+        AlertDialog.Builder builder = new AlertDialog.Builder(requireActivity());
+        LayoutInflater inflater = requireActivity().getLayoutInflater();
 
         View view = inflater.inflate(R.layout.show_shift, null);
 
@@ -97,14 +101,11 @@ public class ShowShift extends AppCompatDialogFragment {
         showShiftdialog.setCancelable(false);
         showShiftdialog.setCanceledOnTouchOutside(false);
 
-        showShiftdialog.setButton(Dialog.BUTTON_NEGATIVE, "OK", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
+        showShiftdialog.setButton(Dialog.BUTTON_NEGATIVE, "OK", (dialog, which) -> {
 
-                showShiftNumber = 0;
+            showShiftNumber = 0;
 //                AttendanceReqUpdate.showShiftNumberUpdate = 0;
-                dialog.dismiss();
-            }
+            dialog.dismiss();
         });
 
         return showShiftdialog;
@@ -283,7 +284,7 @@ public class ShowShift extends AppCompatDialogFragment {
 
         showShiftLists = new ArrayList<>();
 
-        String url = "http://103.56.208.123:8001/apex/ttrams/attendanceUpdateReq/getShiftDetails/"+osm_id+"";
+        String url = api_url_front + "attendanceUpdateReq/getShiftDetails/"+osm_id;
 
         RequestQueue requestQueue = Volley.newRequestQueue(mContext);
 
@@ -311,12 +312,12 @@ public class ShowShift extends AppCompatDialogFragment {
                 updateLay();
             }
             catch (JSONException e) {
-                e.printStackTrace();
+                logger.log(Level.WARNING, e.getMessage(), e);
                 connected = false;
                 updateLay();
             }
         }, error -> {
-            error.printStackTrace();
+            logger.log(Level.WARNING, error.getMessage(), error);
             conn = false;
             connected = false;
             updateLay();
@@ -333,7 +334,7 @@ public class ShowShift extends AppCompatDialogFragment {
                 apptRecyclerView.setAdapter(showShiftAdapter);
             }
             else {
-                AlertDialog dialog = new AlertDialog.Builder(getContext())
+                AlertDialog dialog = new AlertDialog.Builder(mContext)
                         .setMessage("There is a network issue in the server. Please Try later.")
                         .setPositiveButton("Retry", null)
                         .setNegativeButton("Cancel",null)
@@ -355,7 +356,7 @@ public class ShowShift extends AppCompatDialogFragment {
             }
         }
         else {
-            AlertDialog dialog = new AlertDialog.Builder(getContext())
+            AlertDialog dialog = new AlertDialog.Builder(mContext)
                     .setMessage("Please Check Your Internet Connection")
                     .setPositiveButton("Retry", null)
                     .setNegativeButton("Cancel",null)
