@@ -1,7 +1,6 @@
 package ttit.com.shuvo.ikglhrm.leaveAll.leaveBalance;
 
 import android.content.Context;
-import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,8 +12,6 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.ArrayList;
 
 import ttit.com.shuvo.ikglhrm.R;
-import ttit.com.shuvo.ikglhrm.attendance.report.AttenReportAdapter;
-import ttit.com.shuvo.ikglhrm.attendance.report.AttenReportList;
 
 public class LeaveBalanceAdapter extends RecyclerView.Adapter<LeaveBalanceAdapter.LeaveHolder> {
 
@@ -34,8 +31,7 @@ public class LeaveBalanceAdapter extends RecyclerView.Adapter<LeaveBalanceAdapte
     @Override
     public LeaveHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View v = LayoutInflater.from(myContext).inflate(R.layout.leave_balance_list_view, parent, false);
-        LeaveHolder ammvh = new LeaveHolder(v);
-        return ammvh;
+        return new LeaveHolder(v);
     }
 
     @Override
@@ -51,11 +47,6 @@ public class LeaveBalanceAdapter extends RecyclerView.Adapter<LeaveBalanceAdapte
         holder.transfer.setText(aRtl.getTransfered());
         holder.cash.setText(aRtl.getCash_taken());
         holder.balance.setText(aRtl.getBalance_qty());
-
-
-
-
-
     }
 
     @Override
@@ -63,16 +54,16 @@ public class LeaveBalanceAdapter extends RecyclerView.Adapter<LeaveBalanceAdapte
         return leaveBalanceLists.size();
     }
 
-    public class LeaveHolder extends RecyclerView.ViewHolder {
+    public static class LeaveHolder extends RecyclerView.ViewHolder {
 
-        private TextView category;
-        private TextView code;
-        private TextView open;
-        private TextView earn;
-        private TextView consumed;
-        private TextView transfer;
-        private TextView cash;
-        private TextView balance;
+        TextView category;
+        TextView code;
+        TextView open;
+        TextView earn;
+        TextView consumed;
+        TextView transfer;
+        TextView cash;
+        TextView balance;
 
 
         public LeaveHolder(@NonNull View itemView) {
