@@ -1,11 +1,14 @@
 package ttit.com.shuvo.ikglhrm.scheduler;
 
+import android.Manifest;
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Environment;
 
 import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 import androidx.work.Worker;
@@ -27,6 +30,7 @@ import ttit.com.shuvo.ikglhrm.R;
 
 import static android.content.Context.MODE_PRIVATE;
 import static ttit.com.shuvo.ikglhrm.scheduler.Uploader.channelId;
+import static ttit.com.shuvo.ikglhrm.utilities.Constants.api_url_front;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.Request;
@@ -40,7 +44,7 @@ import org.json.JSONObject;
 
 public class MyWorker extends Worker {
 
-//    private String message = null;
+    //    private String message = null;
     private Boolean conn = false;
     private Boolean connected = false;
 
@@ -62,12 +66,13 @@ public class MyWorker extends Worker {
 
     SharedPreferences sharedPreferencesDA;
     public static String FILE_OF_DAILY_ACTIVITY = "";
-    public static  String DISTANCE = "DISTANCE";
-    public static  String TOTAL_TIME = "TOTAL_TIME";
-    public static  String STOPPED_TIME = "STOPPED_TIME";
+    public static String DISTANCE = "DISTANCE";
+    public static String TOTAL_TIME = "TOTAL_TIME";
+    public static String STOPPED_TIME = "STOPPED_TIME";
 
 
     public static final String TASK_DESC_EMP_ID = "task_desc";
+
     public MyWorker(@NonNull Context context, @NonNull WorkerParameters workerParams) {
 
         super(context, workerParams);
@@ -96,7 +101,7 @@ public class MyWorker extends Worker {
         cal.setTime(new Date());
         cal.add(Calendar.DAY_OF_YEAR, -11);
 
-        for (int i = 0 ; i < 10 ;i ++) {
+        for (int i = 0; i < 10; i++) {
             cal.add(Calendar.DAY_OF_YEAR, +1);
             Date calTime = cal.getTime();
             SimpleDateFormat simpleDateFormat = new SimpleDateFormat("dd-MMM-yy", Locale.ENGLISH);
@@ -479,7 +484,7 @@ public class MyWorker extends Worker {
         conn = false;
         connected = false;
 
-        String empDataUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/getEmpData/"+emp_id+"";
+        String empDataUrl = api_url_front + "attendance/getEmpData/" + emp_id;
 
         RequestQueue requestQueue = Volley.newRequestQueue(mContext);
 
@@ -501,19 +506,16 @@ public class MyWorker extends Worker {
                         emp_code = empDataInfo.getString("emp_code");
                     }
                     if (tracking_flag == 1) {
-                        getTrackerUploadDate(job_id[0],coa_id[0],divm_id[0],dept_id[0]);
-                    }
-                    else {
+                        getTrackerUploadDate(job_id[0], coa_id[0], divm_id[0], dept_id[0]);
+                    } else {
                         connected = true;
                         notifyUser();
                     }
-                }
-                else {
+                } else {
                     connected = false;
                     notifyUser();
                 }
-            }
-            catch (JSONException e) {
+            } catch (JSONException e) {
                 connected = false;
                 e.printStackTrace();
                 notifyUser();
@@ -534,7 +536,7 @@ public class MyWorker extends Worker {
         conn = false;
         connected = false;
 
-        String trackerDataUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/getTrackUploadedDate/"+emp_id+"";
+        String trackerDataUrl = api_url_front + "attendance/getTrackUploadedDate/" + emp_id;
 
         RequestQueue requestQueue = Volley.newRequestQueue(mContext);
 
@@ -553,15 +555,14 @@ public class MyWorker extends Worker {
                     }
                 }
                 dateCount = 0;
-                getTrackerDate(job_id,coa_id,divm_id,dept_id);
+                getTrackerDate(job_id, coa_id, divm_id, dept_id);
 
-            }
-            catch (JSONException e) {
+            } catch (JSONException e) {
                 connected = false;
                 e.printStackTrace();
                 notifyUser();
             }
-        },error -> {
+        }, error -> {
             conn = false;
             connected = false;
             error.printStackTrace();
@@ -586,17 +587,17 @@ public class MyWorker extends Worker {
             }
             if (!dateFound) {
                 noDatetoUp = false;
-                String fileName = emp_id+"_"+date+"_track";
+                String fileName = emp_id + "_" + date + "_track";
 
                 FILE_OF_DAILY_ACTIVITY = fileName;
 
                 sharedPreferencesDA = mContext.getSharedPreferences(FILE_OF_DAILY_ACTIVITY, MODE_PRIVATE);
 
-                String dist = sharedPreferencesDA.getString(DISTANCE,null);
-                String totalTime = sharedPreferencesDA.getString(TOTAL_TIME,null);
-                String stoppedTime = sharedPreferencesDA.getString(STOPPED_TIME,null);
+                String dist = sharedPreferencesDA.getString(DISTANCE, null);
+                String totalTime = sharedPreferencesDA.getString(TOTAL_TIME, null);
+                String stoppedTime = sharedPreferencesDA.getString(STOPPED_TIME, null);
 
-                String stringFIle = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS).getPath() + File.separator +  fileName +".gpx";
+                String stringFIle = getApplicationContext().getExternalFilesDir(null).getPath() + File.separator + fileName + ".gpx";
 
                 File blobFile = new File(stringFIle);
 
@@ -611,11 +612,10 @@ public class MyWorker extends Worker {
                     }
                     updatedFiles.add(stringFIle);
                 }
-                uploadEmpTrackerFile(job_id,coa_id,divm_id,dept_id,date,bytes, blobFile,fileName,dist,totalTime,stoppedTime);
+                uploadEmpTrackerFile(job_id, coa_id, divm_id, dept_id, date, bytes, blobFile, fileName, dist, totalTime, stoppedTime);
                 break;
-            }
-            else {
-                System.out.println("Ei "+date +" database e ase");
+            } else {
+                System.out.println("Ei " + date + " database e ase");
             }
         }
         if (noDatetoUp) {
@@ -626,7 +626,7 @@ public class MyWorker extends Worker {
 
     public void uploadEmpTrackerFile(int job_id, String coa_id, String divm_id, String dept_id, String date, byte[] bytes, File blobFile, String fileName, String dist, String totalTime, String stoppedTime) {
 
-        String uploadFileUrl = "http://103.56.208.123:8001/apex/ttrams/attendance/uploadTrackerFile";
+        String uploadFileUrl = api_url_front + "attendance/uploadTrackerFile";
 
         RequestQueue requestQueue = Volley.newRequestQueue(mContext);
 
@@ -638,27 +638,25 @@ public class MyWorker extends Worker {
                 System.out.println(string_out);
                 if (string_out.equals("Successfully Created")) {
                     lastTenDaysFromSQL.add(date);
-                    getTrackerDate(job_id,coa_id,divm_id,dept_id);
-                }
-                else {
+                    getTrackerDate(job_id, coa_id, divm_id, dept_id);
+                } else {
                     System.out.println("EKHANE ASHE 3");
                     connected = false;
                     notifyUser();
                 }
-            }
-            catch (JSONException e) {
+            } catch (JSONException e) {
                 e.printStackTrace();
                 System.out.println("EKHANE ASHE 0");
                 connected = false;
                 notifyUser();
             }
-        },error -> {
+        }, error -> {
             error.printStackTrace();
             System.out.println("EKHANE ASHE -1");
             conn = false;
             connected = false;
             notifyUser();
-        }){
+        }) {
             @Override
             public byte[] getBody() throws AuthFailureError {
                 return bytes;
@@ -668,27 +666,26 @@ public class MyWorker extends Worker {
             public Map<String, String> getHeaders() throws AuthFailureError {
                 Map<String, String> headers = new HashMap<>();
                 if (blobFile.exists()) {
-                    headers.put("P_ELR_ACTIVE","1");
-                    headers.put("P_ELR_FILE_NAME",fileName);
-                    headers.put("P_ELR_MIMETYPE","application/gpx+xml");
-                    headers.put("P_ELR_FILETYPE",".gpx");
+                    headers.put("P_ELR_ACTIVE", "1");
+                    headers.put("P_ELR_FILE_NAME", fileName);
+                    headers.put("P_ELR_MIMETYPE", "application/gpx+xml");
+                    headers.put("P_ELR_FILETYPE", ".gpx");
+                } else {
+                    headers.put("P_ELR_ACTIVE", "0");
+                    headers.put("P_ELR_FILE_NAME", null);
+                    headers.put("P_ELR_MIMETYPE", null);
+                    headers.put("P_ELR_FILETYPE", null);
                 }
-                else {
-                    headers.put("P_ELR_ACTIVE","0");
-                    headers.put("P_ELR_FILE_NAME",null);
-                    headers.put("P_ELR_MIMETYPE",null);
-                    headers.put("P_ELR_FILETYPE",null);
-                }
-                headers.put("P_ELR_EMP_ID",emp_id);
-                headers.put("P_ELR_JOB_ID",String.valueOf(job_id));
-                headers.put("P_ELR_COA_ID",coa_id);
-                headers.put("P_ELR_DIVM_ID",divm_id);
-                headers.put("P_ELR_DEPT_ID",dept_id);
-                headers.put("P_ELR_DATE",date);
-                headers.put("P_ELR_USER",emp_code);
-                headers.put("P_TOTAL_DISTANCE_KM",dist);
-                headers.put("P_TOTAL_TIME",totalTime);
-                headers.put("P_TOTAL_STOPPED_TIME",stoppedTime);
+                headers.put("P_ELR_EMP_ID", emp_id);
+                headers.put("P_ELR_JOB_ID", String.valueOf(job_id));
+                headers.put("P_ELR_COA_ID", coa_id);
+                headers.put("P_ELR_DIVM_ID", divm_id);
+                headers.put("P_ELR_DEPT_ID", dept_id);
+                headers.put("P_ELR_DATE", date);
+                headers.put("P_ELR_USER", emp_code);
+                headers.put("P_TOTAL_DISTANCE_KM", dist);
+                headers.put("P_TOTAL_TIME", totalTime);
+                headers.put("P_TOTAL_STOPPED_TIME", stoppedTime);
                 return headers;
             }
 
@@ -711,9 +708,9 @@ public class MyWorker extends Worker {
 
     public void notifyUser() {
         if (conn) {
-            if(connected) {
+            if (connected) {
                 if (dateCount > 0) {
-                    if (updatedFiles.size() != 0) {
+                    if (!updatedFiles.isEmpty()) {
                         for (int i = 0; i < updatedFiles.size(); i++) {
                             String stringFile = updatedFiles.get(i);
                             File blobFile = new File(stringFile);
@@ -725,12 +722,12 @@ public class MyWorker extends Worker {
                             }
                         }
                     }
-                    if (updatedXml.size() != 0) {
-                        for (int i = 0 ; i < updatedXml.size(); i++) {
+                    if (!updatedXml.isEmpty()) {
+                        for (int i = 0; i < updatedXml.size(); i++) {
 
                             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                                File dir = new File(mContext.getApplicationInfo().dataDir, "shared_prefs/" + updatedXml.get(i)+ ".xml");
-                                if(dir.exists()) {
+                                File dir = new File(mContext.getApplicationInfo().dataDir, "shared_prefs/" + updatedXml.get(i) + ".xml");
+                                if (dir.exists()) {
                                     mContext.getSharedPreferences(updatedXml.get(i), MODE_PRIVATE).edit().clear().apply();
                                     boolean ddd = dir.delete();
                                     System.out.println(ddd);
@@ -741,29 +738,31 @@ public class MyWorker extends Worker {
                         }
                     }
 
+                    NotificationCompat.Builder builder;
+                    NotificationManagerCompat notificationManagerCompat = NotificationManagerCompat.from(mContext);
                     if (dateCount == 1) {
-                        NotificationCompat.Builder builder = new NotificationCompat.Builder(mContext, channelId)
-                                .setSmallIcon(R.drawable.thrn_logo)
+                        builder = new NotificationCompat.Builder(mContext, channelId)
+                                .setSmallIcon(R.drawable.hrm_new_icon_wb)
                                 .setContentTitle("Tracking Service")
                                 .setContentText(dateCount + " File Uploaded")
                                 .setPriority(NotificationCompat.PRIORITY_DEFAULT);
 
-                        NotificationManagerCompat notificationManagerCompat = NotificationManagerCompat.from(mContext);
-//        Uri alarmSound = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+                        //        Uri alarmSound = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
 //        builder.setSound(alarmSound);
-                        notificationManagerCompat.notify(200, builder.build());
                     } else {
-                        NotificationCompat.Builder builder = new NotificationCompat.Builder(mContext, channelId)
-                                .setSmallIcon(R.drawable.thrn_logo)
+                        builder = new NotificationCompat.Builder(mContext, channelId)
+                                .setSmallIcon(R.drawable.hrm_new_icon_wb)
                                 .setContentTitle("Tracking Service")
                                 .setContentText(dateCount + " Files Uploaded")
                                 .setPriority(NotificationCompat.PRIORITY_DEFAULT);
 
-                        NotificationManagerCompat notificationManagerCompat = NotificationManagerCompat.from(mContext);
-//        Uri alarmSound = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+                        //        Uri alarmSound = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
 //        builder.setSound(alarmSound);
-                        notificationManagerCompat.notify(200, builder.build());
                     }
+                    if (ActivityCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.POST_NOTIFICATIONS) != PackageManager.PERMISSION_GRANTED) {
+                        return;
+                    }
+                    notificationManagerCompat.notify(200, builder.build());
 
                 }
                 conn = false;
